### 2022-05-16 #535 @phuongquan
daiquiri: Data quality reporting for temporal datasets
Submitting Author Name: T. Phuong Quan
Submitting Author Github Handle: @phuongquan
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/phuongquan/daiquiri
Version submitted: 0.7.1
Submission type: Standard
Version accepted: TBD
---

```
Package: daiquiri
Title: Data Quality Reporting for Temporal Datasets
Version: 0.7.1
    person(c("T.", "Phuong"), "Quan", email = "phuong.quan@ndm.ox.ac.uk",
        role = c("aut", "cre"), comment = c(ORCID = "0000-0001-8566-1817")),
    person("Jack", "Cregan", role = "ctb"),
    person(family = "University of Oxford", role = "cph"),
    person(family = "National Institute for Health Research (NIHR)", role = "fnd")
    )
Description: Generate reports that enable quick visual review of
    temporal shifts in record-level data. Time series plots showing aggregated
    values are automatically created for each data field (column) depending on its
    contents (e.g. min/max/mean values for numeric data, no. of distinct
    values for categorical data), as well as overviews for missing values,
    non-conformant values, and duplicated rows. The resulting reports are sharable
    and can contribute to forming a transparent record of the entire analysis process.
    It is designed with Electronic Health Records in mind, but can be used for
    any type of record-level temporal data (i.e. tabular data where each row represents
    a single “event”, one column contains the "event date", and other columns
    contain any associated values for the event).
License: GPL (>=3)
    data.table (>= 1.12.8),
    readr (>= 1.3.1),
    ggplot2 (>= 3.1.0),
    scales (>= 1.1.0),
    cowplot (>= 0.9.3),
    reactable (>= 0.2.3),
    stats
    covr,
    codemetar

```

It takes a generic data frame containing raw, record-level, temporal data, and generates a data quality report that enables quick visual review of any unexpected temporal shifts in measures such as missingness, min/max/mean/distinct values, and non-conformance.

The target audience is all researchers who analyse data from large, temporal datasets, particularly routinely-collected data such as electronic health records. The package helps them to quickly check for temporal biases in their data before embarking on their main analyses. It also helps them to do this in a thorough, consistent and transparent way (since the reports are shareable), hence increasing the quality of their studies as well as trust in the scientific process.

To my knowledge, there are a small number of R packages that generate summary statistics and/or data quality reports, (with the two most similar being dataquieR and DQAstats), but none which assist in identifying temporal changes in the data, nor which are as lightweight to use and consume.

Yes

https://github.com/ropensci/software-review/issues/527

Here is the list of items from version 0.0.3.11. The latest development version (0.0.3.13) is not currently working on Windows.

1. These functions do not have examples: [daiquiri] - This is not actually a function, daiquiri is the package help page and includes links to the "most useful" function and vignette.
2. Package uses global assignment operator (‘<<-’). - This is necessary for a withCallingHandlers() call
3. Package has no continuous integration checks. - The package uses GitHub Actions for R-CMD-check and for covr so I don't know why this is showing as a fail.
4. Namespace in Imports field not imported from: ‘reactable’ All declared Imports should be used - The 'reactable' package is in fact used but is probably being missed because it is called from an Rmd file in the inst/rmd folder
5. There are two functions with cyclocomplexity above 15: a) validate_params_type() and b) aggregatefield(). These functions are essentially large switch statements covering a) all arguments to all exported functions (for validating what a user passes to the function, and where some arguments are used across multiple functions), and b) all the different ways that the package can aggregate a data field (e.g. mean, min, max). The only obvious way I can see to reduce the cyclocomplexity score would be to separate out the items into subfunctions, but I judge that would actually make the code harder to follow rather than easier. I am open to other suggestions.
