### 2022-05-16 #535 @mpadge
@phuongquan You're right about the functions not having examples. That was a bug which has now been fixed, so thanks! You can ignore that result. The coverage issue is, however, more problematic. I've tried to run `covr::package_coverage()` (as well as `codecov()` in a couple of different docker containers and always encouter this error:
```
simpleWarning in utils::install.packages(repos = NULL, lib = tmp_lib, pkg$path, : installation of package ‘/daiquiri’ had non-zero exit status

simpleWarning in file(con, "r"): cannot open file '/tmp/RtmpRuQwPt/R_LIBS14f26d8825f2/daiquiri/R/daiquiri': No such file or directory

Error in file(con, "r"): cannot open the connection
```

This is clearly what causes the fails on our system. You should be able to reproduce by cloning your repo in a clean docker container and trying for yourself. I nevertheless see that the GitHub action works, and that your coverage is indeed as you claim, so you may also choose to ignore that if you like. My guess is that it's related to {renv}. `covr` installs the package in a temporary library location, but our check system loads the package prior to that and so populates your `renv` directory.  That then does not get copied across to the temporary libary location used by `covr`, and so installation fails. It works on GitHub actions because that uses a fresh clone of the repo which populates `renv` in the temporary library location.
