### 2022-05-30 #535 @maurolepore
Hi @phuongquan, I'm pleased to be the handling editor of this submission.

I'll re-run checks now. I see the comments above suggest some checks may fail for reasons beyond your control. I'll explore those issues and walk through the [editor's template](https://devdevguide.netlify.app/editortemplate.html#editortemplate) by the end of this week.
