+++ tests/testthat/test_main.R
@@ -1,10 +1,10 @@
 test_that("Test read_data() on csv file created using Excel", {
 	# read file correctly
 	rawdata <- read_data(
-	  test_path("testdata", "specialchars_excel.csv"),
-	  delim = ",",
-	  col_names = TRUE
-	  )
+		test_path("testdata", "specialchars_excel.csv"),
+		delim = ",",
+		col_names = TRUE
+	)
 	expect_equal(nrow(rawdata), 8)
 	expect_equal(length(rawdata), 4)
 
@@ -50,44 +50,49 @@ test_that("create_report() params are present and of correct type", {
 
 test_that("create_report() creates report and returns daiquiri object successfully", {
 	testdf <- read_data(test_path("testdata", "completetestset.csv"))
-	testdaiqobj <- create_report(testdf,
-															 fieldtypes = fieldtypes( col_timepoint_err = ft_ignore(),
-															 												 col_timepoint = ft_timepoint(),
-															 												 col_date_time_err = ft_ignore(),
-															 												 col_date_time = ft_datetime(),
-															 												 col_date_only_err = ft_ignore(),
-															 												 col_date_only = ft_datetime(includes_time = FALSE),
-															 												 col_date_uk_err = ft_ignore(),
-															 												 col_date_uk = ft_datetime(includes_time = FALSE, format = "%d/%m/%Y"),
-															 												 col_id_num_err = ft_ignore(),
-															 												 col_id_num = ft_uniqueidentifier(),
-															 												 col_id_string_err = ft_ignore(),
-															 												 col_id_string = ft_uniqueidentifier(),
-															 												 col_numeric_clean_err = ft_ignore(),
-															 												 col_numeric_clean = ft_numeric(),
-															 												 col_numeric_dirty_err = ft_ignore(),
-															 												 col_numeric_dirty = ft_numeric(),
-															 												 col_categorical_small_err = ft_ignore(),
-															 												 col_categorical_small = ft_categorical(aggregate_by_each_category = TRUE),
-															 												 col_categorical_large_err = ft_ignore(),
-															 												 col_categorical_large = ft_categorical(),
-															 												 col_freetext_err = ft_ignore(),
-															 												 col_freetext = ft_freetext(),
-															 												 col_simple_err = ft_ignore(),
-															 												 col_simple = ft_simple()),
-															 dataset_shortdesc = "completetestset",
-															 aggregation_timeunit = "week",
-															 save_directory = tempdir(),
-															 save_filename = "daiquiri_testthatreport",
-															 log_directory = tempdir(),
-															 showprogress = FALSE)
+	testdaiqobj <- create_report(
+		testdf,
+		fieldtypes = fieldtypes(
+			col_timepoint_err = ft_ignore(),
+			col_timepoint = ft_timepoint(),
+			col_date_time_err = ft_ignore(),
+			col_date_time = ft_datetime(),
+			col_date_only_err = ft_ignore(),
+			col_date_only = ft_datetime(includes_time = FALSE),
+			col_date_uk_err = ft_ignore(),
+			col_date_uk = ft_datetime(includes_time = FALSE, format = "%d/%m/%Y"),
+			col_id_num_err = ft_ignore(),
+			col_id_num = ft_uniqueidentifier(),
+			col_id_string_err = ft_ignore(),
+			col_id_string = ft_uniqueidentifier(),
+			col_numeric_clean_err = ft_ignore(),
[…]
