+++ R/utilities.R
@@ -2,12 +2,12 @@
 
 #' Check all required params have been passed in to the calling function
 #'
-#' Should be called at start of all exported functions to check user has supplied all required arguments
-#' If any are missing, execution is stopped
+#' Should be called at start of all exported functions to check user has
+#' supplied all required arguments If any are missing, execution is stopped
 #'
 #' @param call call from the function being checked
 #' @noRd
-validate_params_required <- function(call){
+validate_params_required <- function(call) {
 	# get the required arguments from function definition
 	params_defined <- formals(as.character(call[[1]]))
 	params_required <- names(which(vapply(params_defined, is.symbol, logical(1))))
@@ -15,124 +15,254 @@ validate_params_required <- function(call){
 	params_passed <- names(as.list(call)[-1])
 
 	if (any(!params_required %in% params_passed)) {
-		stop_custom(.subclass = "invalid_param_missing",
-								message = paste("Required argument(s) missing:",
-																paste(setdiff(params_required, params_passed),
-																			collapse=", ")))
+		stop_custom(
+			.subclass = "invalid_param_missing",
+			message = paste(
+				"Required argument(s) missing:",
+				paste(setdiff(params_required, params_passed),
+							collapse = ", ")
+			)
+		)
 	}
 }
 
-#' Check all params that have been passed in to the calling function are of correct type/class
+#' Check all params that have been passed in to the calling function are of
+#' correct type/class
 #'
-#' Should be called at start of all exported functions to check user has supplied all arguments correctly
-#' Any that are invalid are collated and then execution is stopped
+#' Should be called at start of all exported functions to check user has
+#' supplied all arguments correctly Any that are invalid are collated and then
+#' execution is stopped
 #'
 #' @param call call from the function being checked
-#' @param ... the parameters that were actually passed into the function being checked, with names
+#' @param ... the parameters that were actually passed into the function being
+#'   checked, with names
 #' @noRd
-validate_params_type <- function(call, ...){
+validate_params_type <- function(call, ...) {
 	params_defined <- names(formals(as.character(call[[1]])))
 	params_passed <- list(...)
 	params_names <- names(params_passed)
 
 	# check internal usage is correct
-	if(length(which(params_names != "")) != length(params_passed)){
+	if (length(which(params_names != "")) != length(params_passed)) {
 		stop_custom(.subclass = "invalid_call",
 								message = "Invalid call for function. Params must be passed in with names")
 	}
-	if(!setequal(params_defined, params_names)){
-		stop_custom(.subclass = "invalid_call",
-								message = paste0("Invalid call for function. Different set of params in parent function definition than were passed in to validate_params_type(). In validate_params_type() but not in parent function: ",
-																 paste(setdiff(params_names, params_defined), collapse=", "),
-																 "; In parent function but not in validate_params_type(): ",
-																 paste(setdiff(params_defined, params_names), collapse=", ")))
+	if (!setequal(params_defined, params_names)) {
+		stop_custom(
+			.subclass = "invalid_call",
+			message = paste0("Invalid call for function.
+											 Different set of params in parent function definition than were passed in to validate_params_type().
[…]

[…]
