### 2022-06-03 #535 @maurolepore
Thanks again for this awesome submisison! I'm here to help you pass review smoothly. I looked for opportunities to make the package follow rOpenSci's [guidelines](https://devdevguide.netlify.app/index.html) as closely as possible, and to make the reviewer's job as easy as possible. After the section "Editor checks" you'll see my comments:

* I listed some issues I need to please address before I start searching for reviewers. You may refer to them using the labels ml01, ml02, and so on.

* Also I listed some other issues for you to consider. Those ones have no label and won't block the process
 but if you address them now you're sure they won't be picked up by the reviewers.

Please let me know what questions you have.

---

##### TODO

- [ ] ml01. In README please add a basic example of the package's usage (e.g. [dplyr's usage](https://dplyr.tidyverse.org/index.html#usage)).

Installation instructions:

- [ ] ml02. Please simplify the installation instructions. This code should install daiquiri v0.7.0 and all its dependencies ([details about the `repo` argument](https://remotes.r-lib.org/reference/install_github.html#arguments-1)):

```r
# install.packages("devtools")
devtools::install_github("phuongquan/daiquiri@v0.7.0")
```

Or

```r
# install.packages("pak")
pak::pkg_install("phuongquan/daiquiri@v0.7.0")
```

- [ ] ml03. **Testing**: Please rename for every R/file.R please rename the corresponding file in tests/testhat/ from test_files.R to test-files.R. This change is small and enables the [development workflow with `use_r()` and `use_test()`](https://r-pkgs.org/whole-game.html#use_r).

##### CONSIDER

* [We recommend creating a documentation website for your package using pkgdown](https://devdevguide.netlify.app/building.html#website).

* If you do use pkgdown and rename your vignette from "walkthrough" to "daiquiri" then it will be automatically exposed on the package website under the navigation-bar link as "Get started" (e.g.
[dplyr](https://dplyr.tidyverse.org/articles/dplyr.html), [readr](https://readr.tidyverse.org/articles/readr.html)).

Fit:

* I see a [comment about some ovelap with the pakcages dataquieR and DQAstats](https://github.com/ropensci/software-review/issues/535#issue-1237345405), and I also see some similarities with the packages [skimr](https://docs.ropensci.org/skimr/) and [naniar](https://naniar.njtierney.com/). Consider comparing the alternatives in README or an artile ([exapmle for piggyback](https://docs.ropensci.org/piggyback/articles/alternatives.html)).

Installation instructions:

* It's common to consider the code on GitHub as "in development" and the code on CRAN as "released". If embrace that release model then users installing from GitHub only need the default branch.

* If you stick to using GitHub releases, then consider upgrading them from "pre-release" to "release"  (<https://github.com/phuongquan/daiquiri/releases>).

* The installation instructions are under the heading "Get started". Instead many popular packages use the heading "Installation" (e.g. [dplyr](https://dplyr.tidyverse.org/#installation), [ggplot2](https://ggplot2.tidyverse.org/#installation)), mainy because it's the template you get with
[`usethis::use_readme_rmd()`](https://usethis.r-lib.org/reference/use_readme_rmd.html).

Tests:

* In my local environment `devtools::test()` results in 24 warnings with this message:

```r
semi-transparency is not supported on this device: reported only once per page
```

``` r
devtools::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.2.0 (2022-04-22)
#>  os       Ubuntu 20.04.4 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language (EN)
#>  collate  en_US.UTF-8
#>  ctype    en_US.UTF-8
#>  tz       America/Guatemala
#>  date     2022-06-03
#>  pandoc   2.17.1.1 @ /usr/lib/rstudio/bin/quarto/bin/ (via rmarkdown)
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  brio          1.1.3   2021-11-30 [1] RSPM
#>  cachem        1.0.6   2021-08-19 [1] RSPM
#>  callr         3.7.0   2021-04-20 [1] RSPM
#>  cli           3.3.0   2022-04-25 [1] RSPM
#>  crayon        1.5.1   2022-03-26 [1] RSPM
#>  desc          1.4.1   2022-03-06 [1] RSPM
#>  devtools      2.4.3   2021-11-30 [1] RSPM
#>  digest        0.6.29  2021-12-01 [1] RSPM
#>  ellipsis      0.3.2   2021-04-29 [1] RSPM
#>  evaluate      0.15    2022-02-18 [1] RSPM
#>  fansi         1.0.3   2022-03-24 [1] RSPM
#>  fastmap       1.1.0   2021-01-25 [1] RSPM
#>  fs            1.5.2   2021-12-08 [1] RSPM
#>  glue          1.6.2   2022-02-24 [1] RSPM
#>  highr         0.9     2021-04-16 [1] RSPM
#>  htmltools     0.5.2   2021-08-25 [1] RSPM
#>  knitr         1.39    2022-04-26 [1] RSPM
#>  lifecycle     1.0.1   2021-09-24 [1] RSPM
#>  magrittr      2.0.3   2022-03-30 [1] RSPM
#>  memoise       2.0.1   2021-11-26 [1] RSPM
#>  pillar        1.7.0   2022-02-01 [1] RSPM
#>  pkgbuild      1.3.1   2021-12-20 [1] RSPM
#>  pkgconfig     2.0.3   2019-09-22 [1] RSPM
#>  pkgload       1.2.4   2021-11-30 [1] RSPM
#>  prettyunits   1.1.1   2020-01-24 [1] RSPM
#>  processx      3.5.3   2022-03-25 [1] RSPM
#>  ps            1.7.0   2022-04-23 [1] RSPM
#>  purrr         0.3.4   2020-04-17 [1] RSPM
#>  R.cache       0.15.0  2021-04-30 [1] RSPM
#>  R.methodsS3   1.8.1   2020-08-26 [1] RSPM
#>  R.oo          1.24.0  2020-08-26 [1] RSPM
#>  R.utils       2.11.0  2021-09-26 [1] RSPM
#>  R6            2.5.1   2021-08-19 [1] RSPM
#>  remotes       2.4.2   2021-11-30 [1] RSPM
#>  reprex        2.0.1   2021-08-05 [1] RSPM
#>  rlang         1.0.2   2022-03-04 [1] RSPM
#>  rmarkdown     2.14    2022-04-25 [1] RSPM
#>  rprojroot     2.0.3   2022-04-02 [1] RSPM
#>  rstudioapi    0.13    2020-11-12 [1] RSPM
#>  sessioninfo   1.2.2   2021-12-06 [1] RSPM
#>  stringi       1.7.6   2021-11-29 [1] RSPM
#>  stringr       1.4.0   2019-02-10 [1] RSPM
#>  styler        1.7.0   2022-03-13 [1] RSPM
#>  testthat      3.1.4   2022-04-26 [1] RSPM
#>  tibble        3.1.7   2022-05-03 [1] RSPM (R 4.2.0)
#>  usethis       2.1.6   2022-05-25 [1] RSPM (R 4.2.0)
#>  utf8          1.2.2   2021-07-24 [1] RSPM
#>  vctrs         0.4.1   2022-04-13 [1] RSPM
#>  withr         2.5.0   2022-03-03 [1] RSPM
#>  xfun          0.31    2022-05-10 [1] RSPM (R 4.2.0)
#>  yaml          2.3.5   2022-02-21 [1] RSPM
#>
#>  [1] /home/mauro/R/x86_64-pc-linux-gnu-library/4.2
#>  [2] /opt/R/4.2.0/lib/R/library
#>
#> ──────────────────────────────────────────────────────────────────────────────
```

License:

* Consider running `usethis::use_gpl3_license()`. This is a good way to ensure the license and related infrastructure are setup correctly.

Other:

* Consider deactivating renv with `renv::deactivate()`. I see that [renv can help the develpoer in some ways](https://rstudio.github.io/renv/articles/packages.html) -- such as [to explore issues with backwards and forwards compatibility, {and} reverse dependency checks](https://r-pkgs.org/package-structure-state.html) -- but it seems to add some friction for reviewers. For example, immediatlely after cloning the repo in a fresh project on <https://rstudio.cloud> I get this message:

```
* Project '/cloud/project' loaded. [renv 0.12.2]
Project requested R version '4.1.2' but '4.2.0' is currently being used
* The project may be out of sync -- use `renv::status()` for more details.
```

* Consider running `usethis::use_tidy_style()`. This will style the code in a way that is more likely to be familiar to the reviewers.

* Consider running `spelling::spell_check_package()`.

* Consider running `goodpractice()`. The most frequent suggestion I see is this:

> "avoid long code lines, it's bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters"

```r
# devtools::install_github("mangothecat/goodpractice")
goodpractice::goodpractice()
```

* Consider removing the package reactable from DESCRIPTION or importing one of its functions (maybe via `usethis::use_import_from()`). This should avoid this R CMD check note:

```r
❯ checking dependencies in R code ... NOTE
  Namespace in Imports field not imported from: ‘reactable’
```

* Consider replacing the current package-documentation with the one generated automatically with `usethis::use_package_doc()`. This might help avoid the issue with the bot.

* Consider automating package-development tasks with usethis (https://usethis.r-lib.org/). For a quick tour see <https://r-pkgs.org/whole-game.html>.
