### 2022-06-08 fb1f780
Tidy up examples
