+++ man/daiquiri.Rd
@@ -22,6 +22,6 @@ Classes are S3
 
 The best place to start is the \code{\link{create_report}} function, and the
 walkthrough vignette:
-\href{../doc/walkthrough.html}{\code{vignette("walkthrough", package =
+\href{../doc/daiquiri.html}{\code{vignette("daiquiri", package =
 "daiquiri")}}.
 }
