### 2022-06-08 2b4c8a4
Add example Usage and simplify installation instructions (rOpenSci items ml01/ml03)
