+++ README.md
@@ -66,34 +66,65 @@ conduct this part of the process more thoroughly, consistently and
 transparently, hence increasing the quality of their studies as well as
 trust in the scientific process.
 
-## Getting started
+## Installation
 
 The intention is to make daiquiri available in CRAN but until then, you
 can install the latest release by running the following:
 
 ``` r
+# install dependencies
+install.packages(c("data.table", "readr", "ggplot2", "scales", "cowplot", "rmarkdown", "reactable"))
+
+# install daiquiri with pre-built vignettes
 install.packages("https://github.com/phuongquan/daiquiri/releases/download/v0.7.0/daiquiri_0.7.0.tar.gz", 
                                  repos = NULL, type = "source")
-```
-
-NOTE: Depending on what packages you already have installed, it may
-return an error saying that certain dependencies are not available for
-the package. It will not install these automatically. Once you have
-installed the dependencies from CRAN, try to install the package again.
-
-Once installed, the walkthrough vignette guides you through how to use
-the package:
-
-``` r
-vignette("walkthrough", package = "daiquiri")
 ```
 
 Alternatively, you can install the current development version from
 GitHub (though this may be more buggy):
 
 ``` r
-#install.packages("remotes")
-remotes::install_github("phuongquan/daiquiri", build_vignettes = TRUE)
+#install.packages("devtools")
+devtools::install_github("phuongquan/daiquiri")
+```
+
+## Usage
+
+``` r
+library(daiquiri)
+
+rawdata <- read_data(
+  system.file("extdata", "example_data.csv", package = "daiquiri"),
+  delim = ",",
+  col_names = TRUE
+)
+
+daiqobj <- create_report(
+  rawdata,
+  fieldtypes = fieldtypes(
+    PrescriptionID = ft_uniqueidentifier(),
+    PrescriptionDate = ft_timepoint(),
+    AdmissionDate = ft_datetime(includes_time = FALSE),
+    Drug = ft_freetext(),
+    Dose = ft_numeric(),
+    DoseUnit = ft_categorical(),
+    PatientID = ft_ignore(),
+    Location = ft_categorical(aggregate_by_each_category=TRUE)),
+  override_columnnames = FALSE,
+  na = c("", "NULL"),
+  dataset_shortdesc = "Example data for illustration",
+  aggregation_timeunit = "day",
+  save_directory = ".",
+  save_filename = "example_data_report",
+  showprogress = TRUE,
+  log_directory = NULL
+)
+```
+
+More detailed guidance can be found in the walkthrough vignette:
+
+``` r
+vignette("daiquiri", package = "daiquiri")
 ```
 
 ## How to cite this package
