### 2022-06-08 #535 @phuongquan
Hi @maurolepore,

Many thanks for your comments. I have pushed the following changes:

- ml01: I have added a Usage section to the README
- ml02: I encountered a problem with your suggestion of `devtools::install_github("phuongquan/daiquiri@v0.7.0")` in that it failed for me when I added the option `build_vignettes = TRUE`. I think it might be an issue with `renv ` and I don't know how to fix it, or if other people would get the same error. Since I feel the vignette is very important, I have updated the installation instructions to continue to use the github release but also included a line to explicitly install the dependencies first. Please let me know if this is sufficient. Also, this should no longer be an issue once it's on CRAN
- ml03: I have renamed all the test files to use dashes instead of underscores

Other items:
- I have published a pkgdown site to github pages (https://phuongquan.github.io/daiquiri/index.html)
- I have recreated the Licence using `usethis::use_gpl3_license()`
- In the README, I have renamed the Get Started section to Installation, and moved the reference to the vignette to the Usage section
- I have reformatted much of the code and reduced line lengths where possible

Lastly, The R CMD check note

> Namespace in Imports field not imported from: ‘reactable’ All declared Imports should be used

is erroneous. The 'reactable' package is in fact used (and so needs to be in the DESCRIPTION file) but is probably being missed because it is called from an Rmd file in the inst/rmd folder.

Please let me know if you need anything more from me.

Kind regards,

Phuong
