### 2022-06-10 #535 @maurolepore
Thanks! We are almost there :-)

* ml01. It would be nice to see the output.

- [ ] ml02: I thought about this a bit more and I now think we can't ask the reviewers to consider both the HEAD of the repo and the latest release. If you want them to review the latest release, then please make it the HEAD. Then the simpler installation instructions should be sufficient. After the review you are welcome to again point users to a GitHub release if you still think that's the best idea. In general I think it's good to see what other popular packages do. That will give you a good sense of what most users will expect. Users tend to like packages that are consistent with other packages because they can reuse what they already know, lowering the overall effort they need to use the new package.

- [ ] ml04: Please deactivate renv from the version you submit for review. I thought a bit more about this issue and I now think it's imprtant to address it before review. Reviewers will have to deactivate renv to access packages they need for the review (e.g. devtools) and renv blocks that. Please be kind with the voluntary reviewers and do it yourself. After the review you are welcome to activate renv again.

- [ ] ml05: Please ensure the examples for `create_report()` and `report_data()` run in 5s or less. This is required for CRAN and kind to the reviewers who may run R CMD check multiple times. Also it may distract them from other more interesting issues. R CMD check shows this:

```r
✔  checking examples (36.8s)
   Examples with CPU (user + system) or elapsed time > 5s
                   user system elapsed
   create_report 17.695  0.679  18.859
   report_data   13.871  0.165  14.737
```

- [ ] Please address the NOTE about reactable.  You can do it by moving "reactable" from Imports: to Suggests of by using any of its functions in any file in the R/ directory. If you use a function only for this purpose please write a comment explaining so -- else the reviewers might wonder what's going on. I want to avoid this relatively simple NOTE to distract reviewers from more interesting issues. Anyway you'll need to address it before you submit to CRAN.

```r
❯ checking dependencies in R code ... NOTE
  Namespace in Imports field not imported from: ‘reactable’
    All declared Imports should be used.
```
