### 2022-06-15 0f146a0
Addresses R CMD Check NOTE: Namespace in Imports field not imported from: ‘reactable’ (rOpenSci item ml06)
