### 2022-06-15 4af30f0
Wrap long-running examples in \donttest (rOpenSci item ml05) and update examples
