+++ R/main.R
@@ -66,12 +66,15 @@ NULL
 #'   character will still be processed according to the fieldtypes
 #'   specification.
 #' @examples
+#' \donttest{
+#' # load example data into a data.frame
 #' rawdata <- read_data(
 #'   system.file("extdata", "example_data.csv", package = "daiquiri"),
 #'   delim = ",",
 #'   col_names = TRUE
 #' )
 #'
+#' # create a report in the current directory
 #' daiqobj <- create_report(
 #'   rawdata,
 #'   fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
@@ -84,6 +87,7 @@ NULL
 #'     Location = ft_categorical(aggregate_by_each_category=TRUE)),
 #'   override_columnnames = FALSE,
 #'   na = c("","NULL"),
+#'   dataset_shortdesc = "Example data provided with package",
 #'   aggregation_timeunit = "day",
 #'   save_directory = ".",
 #'   save_filename = "example_data_report",
@@ -92,6 +96,7 @@ NULL
 #' )
 #'
 #' \dontshow{file.remove("./example_data_report.html")}
+#' }
 #'
 #' @seealso \code{\link{read_data}}, \code{\link{fieldtypes}},
 #'   \code{\link{availablefieldtypes}}
