+++ man/report_data.Rd
@@ -38,12 +38,15 @@ A string containing the name and path of the saved report
 Generate report from previously-created sourcedata and aggregatedata objects
 }
 \examples{
+\donttest{
+# load example data into a data.frame
 rawdata <- read_data(
   system.file("extdata", "example_data.csv", package = "daiquiri"),
   delim = ",",
   col_names = TRUE
 )
 
+# validate and prepare the data for aggregation
 sourcedataobj <- prepare_data(
   rawdata,
   fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
@@ -56,15 +59,18 @@ sourcedataobj <- prepare_data(
     Location = ft_categorical(aggregate_by_each_category=TRUE)),
   override_columnnames = FALSE,
   na = c("","NULL"),
+  dataset_shortdesc = "Example data provided with package",
   showprogress = TRUE
 )
 
+# aggregate the data
 aggregatedataobj <- aggregate_data(
   sourcedataobj,
   aggregation_timeunit = "day",
   showprogress = TRUE
 )
 
+# save a report in the current directory using the previously-created objects
 report_data(
   sourcedataobj,
   aggregatedataobj,
@@ -74,6 +80,7 @@ report_data(
 )
 
 \dontshow{file.remove("./example_data_report.html")}
+}
 
 }
 \seealso{
