### 2022-06-15 ed83749
Simplify installation instructions (rOpenSci item ml02)
