### 2022-06-15 94d1f70
Add example report to github pages
