### 2022-06-15 a07133d
Fix outdated note about column names
