### 2022-06-15 #535 @phuongquan
Hi @maurolepore,

I have pushed the following changes:

- ml01: I have added an example report to the package website and added a link to it in the Usage section of the README
- ml02: I want the reviewers to consider only the HEAD of the repo so have removed the reference to the latest 'release' from the installation instructions
- ml04: I have deactivated renv
- ml05: I have wrapped the long running examples in /donttest and they no longer appear in R CMD check. (I can't make them run any quicker as it is the rmarkdown rendering that is taking the time)
- ml06: I have added a dummy function to stop the NOTE about reactable

Please let me know if you need anything more.

Thanks,
Phuong
