### 2022-06-16 #535 @maurolepore
Thanks! I'll run checks one last time and start searching for reviewers shortly.

Could you please name three people you think might be good reviewers and explain why?

(I would appoint only one but having more helps understand what kinds of skill you think are useful.)

Here are some more non-blocking comments.

* ml01. Maybe you could minimize your example to somethis like this:

```r
library(daiquiri)

path <- daiquiri_example("raw_data.csv")
raw_data <- readr::read_csv(path)
raw_data

if (interactive()) {
  create_report(raw_data)
}
```
Here is a version comments for you

```r
library(daiquiri)

# Maybe write this helper (inspired by `readr::readr_example`)
path <- daiquiri_example("raw_data.csv")
# Maybe create a basic .csv that reads well directly with read_csv()?
# I think read_data() might be helpful but not the main goal of the package
raw_data <- readr::read_csv(path)

# Do show what the data looks like. Users may want to know if the data they
# have is a good fit for this package. Ensure to show just a bit. If it doesn't print
# as a tibble, then use `head()` or `str()`.
raw_data

# It won't run when you render README. It  will run if copy-pasting this code
if (interactive()) {
  # If you add this default: `fieldtypes = fieldtypes()`
  create_report(raw_data)
}
```

* ml05: Would it run faster if you make the example raw_data smaller, like 1/4 or 1/10 of the size it has now?
