### 2022-06-17 3e6f817
Fixed error when user passes in a data.table with non-character columns, and explicitly take a copy of the df earlier
