+++ R/sourcedata.R
@@ -77,7 +77,7 @@ is.datafield <- function(x) inherits(x, "datafield")
 #'   \code{\link{aggregate_data}}, \code{\link{report_data}},
 #'   \code{\link{create_report}}
 #' @export
-#' @importFrom data.table .N
+#' @importFrom data.table .N .SD
 prepare_data <- function(df,
 												 fieldtypes,
 												 override_columnnames = FALSE,
@@ -126,33 +126,6 @@ prepare_data <- function(df,
 	validate_columnnames(names(df),
 											 names(fieldtypes),
 											 check_length_only = override_columnnames)
-	if (override_columnnames == TRUE) {
-		names(df) <- names(fieldtypes)
-	}
-
-	# ensure all columns are character type because readr::type_convert won't skip numeric columns
-	df_datatypes <- vapply(df, typeof, character(1))
-	df_nonchar_warnings <- data.table::data.table()
-	if (any(df_datatypes != "character")) {
-		# Report presence of any non-char columns in source data frame (except ignored ones)
-		df_nonchar_warnings <-
-			data.table::data.table(
-				colindex = which(
-					df_datatypes != "character" &
-						!vapply(fieldtypes, is.fieldtype_ignore, logical(1))
-				),
-				rowindex = NA,
-				message = paste0(
-					"Data supplied as ",
-					df_datatypes[which(df_datatypes != "character" &
-														 	!vapply(fieldtypes, is.fieldtype_ignore, logical(1)))],
-					" instead of character, non-conformant values will not be identified"
-				)
-			)
-		# update the df
-		df[df_datatypes != "character"] <-
-			lapply(df[df_datatypes != "character"], as.character)
-	}
 
 	log_message(paste0("Importing source data [", dataset_shortdesc, "]..."),
 							showprogress)
@@ -161,27 +134,57 @@ prepare_data <- function(df,
 	rows_source_n <- nrow(df)
 	# number of columns in source
 	cols_source_n <- length(df)
-	timepoint_index <- which(vapply(fieldtypes, is.fieldtype_timepoint, logical(1)))
-	timepoint_fieldname <- names(timepoint_index)
+
+	# take a copy of the df so that their original doesn't get updated unknowingly
+	# and convert df to data.table to ensure consistency hereonwards whether user
+	# passes in a data.frame or a data.table
+	# TODO: consider removing df at this point, to release memory
+	dt <- data.table::as.data.table(df)
+
+	if (override_columnnames == TRUE) {
+		names(dt) <- names(fieldtypes)
+	}
 
 	# Validate data against specification, store warnings instead of printing them
 	# use readr::type_convert for now.  Ideally want to store original values and describe action taken too
-	# convert to data.table at this point
+
+	# ensure all columns are character type because readr::type_convert won't skip numeric columns
+	dt_datatypes <- vapply(dt, typeof, character(1))
+	dt_nonchar_warnings <- data.table::data.table()
+	if (any(dt_datatypes != "character")) {
+		# Report presence of any non-char columns in source data frame (except ignored ones)
+		dt_nonchar_warnings <-
+			data.table::data.table(
+				colindex = which(
+					dt_datatypes != "character" &
+						!vapply(fieldtypes, is.fieldtype_ignore, logical(1))
+				),
+				rowindex = NA,
+				message = paste0(
+					"Data supplied as ",
+					dt_datatypes[which(dt_datatypes != "character" &
+														 	!vapply(fieldtypes, is.fieldtype_ignore, logical(1)))],
[…]
