+++ NAMESPACE
@@ -25,3 +25,4 @@ export(report_data)
 importFrom(data.table,':=')
 importFrom(data.table,.EACHI)
 importFrom(data.table,.N)
+importFrom(data.table,.SD)
