### 2022-06-17 c5cc1a1
Fixed bug (introduced in d779fd1a) where ignored fields were appearing in wrong tab in report
