### 2022-06-23 #535 @phuongquan
Hi @maurolepore,

I'd like to suggest the following reviewers:

These two people have worked on R packages looking at data quality in the same research field that daiquiri was originally developed for:

- Stephan Struckmann <stephan.struckmann@uni-greifswald.de> maintainer of "dataquieR: Data Quality in Epidemiological Research"
- Hossein Estiri <hestiri@mgh.harvard.edu> author of "DQe-v: A Database-Agnostic Framework for Exploring Variability in Electronic Health Record Data Across Time and Site Location" https://doi.org/10.13063%2F2327-9214.1277

And this person comes from a more general area of reviewing data in a friendly/attractive way:

- Agustin Calatroni <agustin_calatroni@rhoworld.com> maintainer of "dataxray: An interactive table interface for data summaries"

Thanks,
Phuong
