### 2022-06-24 #535 @phuongquan
Hi @maurolepore,

I just read the Peer review policies and think the suggestions I made may have conflicts of interest in that they are significant contributors to what may be seen as "competitor projects". I would like to suggest the following reviewers instead, who are more from the user side:

- Rahul Sangole [@rsangole] who is a data scientist interested in time series and anomaly detection, and who has starred the github project
- Daniel T. Holmes (https://www.providencelaboratory.com/staff_holmes.php), who is the corresponding author of "Reproducible Research and Reports with R" https://doi.org/10.1373/jalm.2018.028373
- Brad Cannell (https://www.bradcannell.com/) who is an epidemiologist and author of the ebook "R for Epidemiology" https://www.r4epi.com/

Thanks!
Phuong
