### 2022-06-29 18f76a1
Minimise example and show table head (rOpenSci item ml01)
