### 2022-06-29 08f2764
Make it clearer that the main purpose of read_data() is to read in the file without doing any datatype conversion
