+++ vignettes/daiquiri.Rmd
@@ -45,11 +45,11 @@ There are some missing values which are represented by the string 'NULL'. String
 library(daiquiri)
 
 # this is where the example file is located
-filename <- system.file("extdata", "example_data.csv", package = "daiquiri")
+path <- system.file("extdata", "example_data.csv", package = "daiquiri")
 
-# load the data into a data.frame
+# load the data into a data.frame without doing any datatype conversion
 example_data <- read_data(
-  filename,
+  path,
   delim = ",",
   col_names = TRUE,
   showprogress = FALSE
@@ -104,6 +104,9 @@ fts <- fieldtypes(
 ```
 
 This is just one of many different possible choices, for example another option could be to use the AdmissionDate as the timepoint field instead, and set the Drug column to be categorical. It is up to you to decide which set of time series you want to be created for each column.
+
+TIP: If you have a lot of columns in your dataset you can use the `fieldtypes_template()` function to print a template specification to the console that you can copy and edit.
+
 
 ## Generating a data quality report
 
