### 2022-07-22 #535 @phuongquan
Hi @maurolepore, I've noticed that the label is still showing as 1/editor-checks when it should be 2/seeking-reviewers. I imagine something went wrong with the bot? I just wanted to raise it in case it makes a difference in moving the process along.

Thanks,
Phuong
