### 2022-07-22 #535 @maurolepore
Thanks! You're right. I just changed the label manually.

FYI I'm still in search of the second reviewer. Sometimes it's quick; sometimes it's not so quick. Once we reach out to someone we give them a few days to respond before we move to someone else.
