### 2022-07-22 #535 @rsangole
I really appreciate the request to review, and I'd love to do so on this tool too. Unfortunately, Apple (where I work) is pretty strict about open source contributions, so it could be months before I get approval to do so. Your timeline may benefit finding another reviewer.
