### 2022-08-08 #535 @mbcann01
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

Thank you for the opportunity to review this package. It looks really cool! I don't personally use time series data that often, but I can see how this package would be useful if I did. I have a couple of comments below that I hope will be helpful to the authors.

-   The example in the "why should I use it" section of README. The figures are nice, but I'm not 100% sure what the connections I'm supposed to be making. Perhaps showing the Y-axis would help me to interpret the figures. Some users may even appreciate you spelling it out for them. Something like, "in the first figure... As you can see, that is reflected in the second figure as...".

-   Example data. This is certainly optional, but I've started including example data in my packages. In the README, it looks as though you are including an example csv file in inst/extdata. To make testing/following along even easier for your users, you may want to consider converting this to exported data that users can load with `data(daiquiri_data)` or something like that. Here are some instructions if you are interested: <https://r-pkgs.org/data.html#sec-data-data>

-   Vignette. I appreciated the details in the vignette!

-   I didn't see instructions for contributing anywhere on the README. I apologize if I just overlooked them.

-   All of the examples shown on README ran without issues on my local machine.
