### 2022-09-01 #535 @maurolepore
Dear @phuongquan,

I discussed with the editors board and I think it would be best to either a) find a new reviewer or b) review the the package myself. Which one do you prefer?

My review would be faster and I substantial experience building packages in general. But an external reviewer has greater potential for noticing details specific to your field.

It's very rare for @elinw to be unresponsive, so likely there is a good reason for it, and seems best to release the pressure from her.
