### 2022-09-13 #535 @phuongquan
Hi @maurolepore, I just wanted to check if you were waiting for me to respond to the check results or if I should be waiting for you to post a full review?

Thanks
