### 2022-09-18 #535 @maurolepore
Thanks a lot for your patience. I'm pleased to share my review. Please take my comments in bullets as suggestions, and check-boxes as requests. In any case, feel free to justify and not follow my advice. All my comments are prefixed with "r2" (reviewer 2) and numbered, to make it easy for you to respond to them.

I had no working relationship with the author. I am the assigned editor, now in the
reviewer role because one  appointed reviewer is now unavailable.

	* [ ] (r02-01): In the vignette consider removing the literal reference to
	"NOTE". The text itself is noteworthy and the announcement "NOTE" can be
	dropped and yet be clear. Similarly, the headings "Example:" seems redundant
	when it's followed by the text "For example, ...". You may drop the heading as
	the sections are short and read well without subdivision.

    - [ ] (r2-02): Please ensure to document the return value of every function,
    even it is `NULL`. CRAN enforces it. For example fix this in `log_close()`.
    - [ ] (r2-03): Ensure to format `code` and [links](www.google.com) as such.
    With roxygen2 you do this with backticks and square brackets, respectively.
    For example, the documentation of `fieldtypes_template()` mentions
    "fieldtype()" -- which is not formatted as code nor linked to its helpfile.
    - [ ] (r2-04): The package documentation typically lives in R/packagename.R
    not R/main.R. See `?usethis::use_package_doc()` and
    `?usethis::use_import_from()`.

	- [ ] (r2-05): Consider using more of the convenient features  of roxygen2,
	e.g. use [some_link] instead of \\link{some_link} and `some_code` instead of
	\\code{some_code}.

  * (r2-06): Ensure the use of `\donttest{}`, `\dontrun{}`, and `\dontshow()` as
  per CRAN policies. They are common sources of rejection.

  - [ ] (r2-07): Whenever possible, print the result of the examples. For
  example the result of `fileldtypes()` is assigned to `fts` but not printed.
  Printed results help the reader see the output without the need of running the
  code themselves. Also please show one example of a `ft_*()` function (e.g.
  `fs_simple()`) so the reader can see its structure.

- [x] **Community guidelines:** including contribution guidelines in the README
or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`
(which may be autogenerated via `Authors@R`).

	- [ ] (r2-08): Performance might be an issue. I see tests and examples that
	take a long time to run. Consider using smaller data or removing performance
	bottlenecks.

	- [ ] (r2-09): I fail to run test-sourcedata.R -- it takes too long or freezes.
	After removing test-sourcedata.R I successfully run all tests but they still
	take too long. The slower your tests, the less frequently you'll run them, and
	the more your code becomes prone to errors. For companison run the tests
	of other popular packages on CRAN.

  ```
	ℹ Testing daiquiri
	✔ | F W S  OK | Context
	✔ |       133 | aggregatedata [13.4s]
	✔ |        15 | fieldtypes [1.0s]
	✔ |        24 | main [240.8s]
	✔ |         7 | reports [110.8s]
	✔ |        45 | utilities [1.4s]
	```

	- [ ] (r2-10) Adapt the tests or error-class so that each test is sensitive to
	a specific error. I see a number of cases where different errors are tested
	against the same class and no other property of the error.

	- [ ] (r2-11): In general, try to use one test file for each function you  test
	(e.g. usually each exported function). This plays well with development
	functions in the usethis package like `use_test()`, `use_test()`, and
	`rename_files()`. Because testthat automatically prints the file name in the
	test report, this approach allows you to not type the name of the function
	inside the call to `test_that()` -- making tests easier to write and maintain.

	```r
	# good
	# tests/testthat/test-f.R
	test_that("works as expected", { code })

	# bad
	# tests/testthat/test-main.R
	test_that("f() works as expected", { code })
	```

	- [ ] (r2-12): Make your tests granular enough so that a specific failure would
	point to a specific expectation, and thus a specific piece of code. This will
	make debugging easier. In `validate_columnnames()` I see a single call to
	`test_that()` has multiple expectations, some expecting good behavior and
	others expecting errors; they will be more informative if those different
	expectations live in different tests.

	- [ ] (r2-13): Please resolve the items marked with "X" by the command
	`@ropensci-review-bot check package` (see
	https://github.com/ropensci/software-review/issues/535#issuecomment-1236161582).
	The section Editor-in-chief instructions state "Processing may not proceed
	until the items marked with ✖️ have been resolved."

	- [ ] (r2-14): Please note all results by the command `@ropensci-review-bot
	check package` and look for opportunities to improve the package. For example,
	the goodpractice section reports that many lines are longer than the
	recommended width of 80 characters. You may reflow lines relatively easy (and
	semi-automatically) and improve readability by a lot.

	- [ ] (r2-15): See rOpenSci's guidelines about [code
	style](https://devguide.ropensci.org/building.html?q=tidyverse%20style#code-style).
	Using a style that is widely used in R and particularly in rOpenSci would be
	more friendly for contributors.

	> For more information on how to style your code, name functions, and R scripts
	inside the R/ folder, we recommend reading the code chapter in The R Packages
	book. We recommend the styler package for automating part of the code styling.
	We suggest reading the Tidyverse style guide.

	For example, please use snake case consistently everywhere -- including the
	names of functions, files, arguments, and variables:

	```
	INCONSISTENT            CONSISTENT
	override_columnnames -> override_column_names
	showprogress         -> show_progress
	save_directory       -> save_directory
	```

	* (r2-17): See rOpenSci's guidelines about [naming your
	package](https://devguide.ropensci.org/building.html#naming-your-package). In a
	Google seardch the name "daiquiri" seems be too common to help discover the
	package.

	- [ ] (r2-18): Ensure exported functions use the format `verb_noun()`
	consistently. For example I see this format in `create_report()` but not in
	`log_close()`.

	- [ ] (r2-19): Use related prefixes consistently. For example `fieldtype()` and
	`fieldtype_template()` are consistent, but ``ft_timepoint()` is not. (But note
	that `field_type*()` is better because it's also consistent with other names
	using snake case.)

Estimated hours spent reviewing: 10h

> Please do not list editors as contributors. Your participation in and
contribution to rOpenSci is thanks enough!
-- https://devguide.ropensci.org/collaboration.html?q=acknowledgements#attributions

---

* (r2-20) Recently I have seen a push towards quarto. Consider the pros and cons
of quarto versus rmarkdown.

- [ ] (r2-21): Rename the example dataset more specifically to preserve the
option of adding more datasets in the future. For example, if later you have two
datasets, the names `example_dataset.csv` and `example_dataset2.csv` are less
descriptive than `prescriptions.csv` and `surgeries.csv`.

- [ ] (r2-22): Remove unused code. For example, I see commented-out code in
tests/testthat/test-utilities.R and R/changepoints.R. Should you need that code
later, you can find it with `git log -S "some code"`. And if you plan to work
with that code soon, it may be best to move it to a GitHub issue and follow up
in a GitHub pull requests when you are ready.

- [ ] (r2-23): Consider removing the function `read_data()`. It seems to
be a thin wrapper around `readr::read_delim()`. If you keep it please do the
following:

    1. Explain why users should not read the data with `readr::read_delim()` and
    friends directly.
    2. Consider preserving the name of the arguments to `readr::read_delim()`
    and their original defaults to avoid confusing users of that popuolar
    function.
    3. Allow users to pass arguments to `readr::read_delim()` via `...`.

* (r2-24): Consider printing objects of class "fieldtypes" as a tibble. That
would be easy for the reader to understand and for you to maintain (instead of
creating a new print method). For example, this is what this idea could look
like:

	```r
	fieldtypes(
		PrescriptionID = ft_uniqueidentifier(),
		PrescriptionDate = ft_timepoint()
	)
	#> # A tibble: 2 × 3
	#>   column           field_type        options
	#>   <chr>            <chr>             <chr>
	#> 1 PrescriptionID   unique identifier <NA>
	#> 2 PrescriptionDate time point        includes time
	```

- [ ] (r2-25): For a first release consider exporting the minimum code possible.
For example: a) `fieldtypes()`, b) all the `ft_*()` functions, c)
`prepare_data()`, d) `aggregate_data()`, and e) `report_data()`. Here are some
benefits:

	* The surface for bugs is smaller.
	* Less documentation.
	* Fewer features to test, so you can test them more intensively.
	* More likely the package will pass the first, human inspected, CRAN
	submission. The first CRAN submission is the hardest because it's the only
	submission that gets inspected by a human. The more code you have, the more
	issues CRAN might find. All other submissions skip human inspection -- unless
	the automated checks detect some problem.
	* Learn from the first CRAN submission before you work on a bigger submission.
	You will learn from the rigor that CRAN enforces, and from the feedback you'll
	gather from the first users of the package. It's always easier to add features
	than it is to remove them ( see [Package evolution - changing stuff in your
	package](https://devguide.ropensci.org/evolution.html?q=life#the-lifecycle-package)).

- [ ] (r2-26): In `report_data()` allow users to pass arguments to
`rmarkdown::render()` via the `...` argument.

- [ ] (r2-27): In `prepare_data()`, `aggregate_data()`, and `report_data()`,
print messages with dedicated functions like `message()` or `warning()` -- not
with `print()`, `cat()` or similar. This allows users to handle messages with
condition-handling features like `tryCatch()`, `suppressMessages()`, and
friends. Also, the primary goal of these functions is to compute something, so
printing a message is only a side effect. Thus by default messages should likely
be quiet. See [Side effect
soup](https://design.tidyverse.org/side-effect-soup.html#what-is-a-side-effect).

- [ ] (r2-28): Consider handling missing arguments using R's default. The custom
code for this seems to add complexity and maintenance code with relatively
little value.

``` r
# Good enough?
f <- function(required) {
  required
}

f()
#> Error in f(): argument "required" is missing, with no default

# Overly complicated?
g <- function(required) {
  validate_params_required(match.call())
  required
}

g()
#> Error in validate_params_required(match.call()): could not find function "validate_params_required"
```

- (r2-29): I see a heavy use of custom classes. In R this is relatively
uncommon. Ensure their value justify the complexity they add. Consider using
the vctrs package (https://vctrs.r-lib.org/articles/s3-vector.html).

- (r2-30): I see deep nested logic (see Jenny Bryan's [Code smells and
feels](https://github.com/jennybc/code-smells-and-feels#readme)). In the
goodpractice section of the rOpenSci's bot checks I see that the following
functions have cyclocomplexity >= 15:

```
function               cyclocomplexity
validate_params_type   84
aggregatefield         56
```

- (r2-31): In `report_data()` I see a call to `package = utils::packageName()`.
Please ensure to test this works as you expect when the user calls the function
from anywhere -- particularly from outside the source of the daiquiry package.

* (r2-31): Ensure the code complies with the general principles in the Mozilla
reviewing guide.
