### 2022-10-04 8b2c981
Make Readme figures clearer. rOpenSci review item @mbcann01
