### 2022-10-04 9ee0373
Rename exported functions to verb_noun format. rOpenSci item (r2-18)
