+++ man/print_fieldtypes_template.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/fieldtypes.R
-\name{fieldtypes_template}
-\alias{fieldtypes_template}
+\name{print_fieldtypes_template}
+\alias{print_fieldtypes_template}
 \title{Print a template fieldtypes specification to console}
 \usage{
-fieldtypes_template(df, default_fieldtype = ft_ignore())
+print_fieldtypes_template(df, default_fieldtype = ft_ignore())
 }
 \arguments{
 \item{df}{data frame including the column names for the template
@@ -23,7 +23,7 @@ or \code{\link{prepare_data}}().
 \examples{
 df <- data.frame(col1 = rep("2022-01-01", 5), col2 = rep(1, 5), col3 = 1:5, col4 = rnorm(5))
 
-fieldtypes_template(df, default_fieldtype = ft_numeric())
+print_fieldtypes_template(df, default_fieldtype = ft_numeric())
 }
 \seealso{
 \code{\link{fieldtypes}}
