+++ man/close_log.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/utilities.R
-\name{log_close}
-\alias{log_close}
+\name{close_log}
+\alias{close_log}
 \title{Closes any active log file}
 \usage{
-log_close()
+close_log()
 }
 \description{
 Closes any active log file
 }
 \examples{
-log_close()
+close_log()
 }
