+++ man/initialise_log.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/utilities.R
-\name{log_initialise}
-\alias{log_initialise}
+\name{initialise_log}
+\alias{initialise_log}
 \title{Initialise log file}
 \usage{
-log_initialise(log_directory)
+initialise_log(log_directory)
 }
 \arguments{
 \item{log_directory}{String containing directory to save log file}
@@ -17,10 +17,10 @@ Choose a directory in which to save the log file. If this is not called, no
 log file is created.
 }
 \examples{
-  logname <- log_initialise(".")
+  logname <- initialise_log(".")
 
   \dontshow{
-  log_close()
+  close_log()
   file.remove(logname)
   }
 }
