+++ NAMESPACE
@@ -5,10 +5,10 @@ S3method(print,daiquiri_object)
 S3method(print,fieldtypes)
 S3method(print,sourcedata)
 export(aggregate_data)
+export(close_log)
 export(create_report)
 export(export_aggregated_data)
 export(fieldtypes)
-export(fieldtypes_template)
 export(ft_categorical)
 export(ft_datetime)
 export(ft_freetext)
@@ -17,9 +17,9 @@ export(ft_numeric)
 export(ft_simple)
 export(ft_timepoint)
 export(ft_uniqueidentifier)
-export(log_close)
-export(log_initialise)
+export(initialise_log)
 export(prepare_data)
+export(print_fieldtypes_template)
 export(read_data)
 export(report_data)
 importFrom(data.table,':=')
