### 2022-10-04 52d302b
Rename exported functions to verb_noun format (cont.). rOpenSci item (r2-18)
