+++ tests/testthat/test-fieldtypes.R
@@ -56,20 +56,20 @@ test_that("Fieldtype colnames cannot include reserved words", {
 })
 
 
-test_that("fieldtypes_template() params are present and of correct type", {
-	expect_error(fieldtypes_template(),
+test_that("print_fieldtypes_template() params are present and of correct type", {
+	expect_error(print_fieldtypes_template(),
 							 class = "invalid_param_missing")
 
-	expect_error(fieldtypes_template(df = c("Fieldname", 123)),
+	expect_error(print_fieldtypes_template(df = c("Fieldname", 123)),
 							 class = "invalid_param_type")
 
-	expect_error(fieldtypes_template(df = data.frame("Fieldname" = 123),
+	expect_error(print_fieldtypes_template(df = data.frame("Fieldname" = 123),
 														default_fieldtype = TRUE),
 							 class = "invalid_param_type")
 })
 
-test_that("fieldtypes_template() generates template fieldtypes output", {
-	expect_snapshot_output(fieldtypes_template(df = data.frame("col1" = 123,
+test_that("print_fieldtypes_template() generates template fieldtypes output", {
+	expect_snapshot_output(print_fieldtypes_template(df = data.frame("col1" = 123,
 																														 "col2" = 123,
 																														 "col3" = "hello")))
 })
