+++ tests/testthat/_snaps/fieldtypes.md
@@ -1,4 +1,4 @@
-# fieldtypes_template() generates template fieldtypes output
+# print_fieldtypes_template() generates template fieldtypes output
 
     fieldtypes( "col1" = ft_ignore(),
     	"col2" = ft_ignore(),
