### 2022-10-05 fc55cd9
Remove unused code - changepoints functionality. rOpenSci item (r2-22)
