+++ R/changepoints.R
@@ -1,301 +1,0 @@
-# NOTE: Changepoints functionality disabled until we find a method that works
-
-# # Code for creation of changepoints object
-# # these contain both the (vector) data for the changepoints and the relevant metadata
-# # they should be added as children of aggregatefield objects
-# #
-#
-# # -----------------------------------------------------------------------------
-# # all changepointresults for an individual aggregatefield
-# find_changepoints <- function(aggfieldvalues, method = "all", showprogress = FALSE){
-#   #temp assignment
-#   #   aggfieldvalues<-outpatdata_byday$aggregatefields[[4]]$values
-#   # method = "all"
-#
-# 	if( method == "none" ){
-# 		log_message(paste0("Skipping changepoints calculation"), showprogress)
-# 		changepoints <- vector("list", 0)
-# 	}
-# 	else{
-# 		log_message(paste0("Calculating changepoints..."), showprogress)
-# 		aggtypes = names(aggfieldvalues)[-1]
-#
-# 	  # # flat list
-# 	  # # initialise list to bigger than needed
-# 	  # changepoints <- vector("list", length(aggtypes)*4)
-# 	  # index <- 1
-# 	  # for(i in seq_along(aggtypes) ){
-# 	  #   #    i = 1
-# 	  #   # zeros are not meaningful for dates
-# 	  #   if( aggtypes[i] %in% c("n","missing_n","missing_perc","distinct","min","max") &&
-# 	  #       (!(inherits(aggfieldvalues[[aggtypes[i]]], "POSIXct") || inherits(aggfieldvalues[[aggtypes[i]]], "Date"))) ){
-# 	  #     changepoints[[index]] <- changepoints_is_zero(aggfieldvalues, aggtype = aggtypes[i], nainterpretation = "ignore")
-# 	  #     names(changepoints)[index] <- paste0(aggtypes[i], "__is_zero")
-# 	  #     index <- index + 1
-# 	  #     }
-# 	  #   if( aggtypes[i] != "n" ){
-# 	  #     changepoints[[index]] <- changepoints_is_na(aggfieldvalues, aggtype = aggtypes[i])
-# 	  #     names(changepoints)[index] <- paste0(aggtypes[i], "__is_na")
-# 	  #     index <- index + 1
-# 	  #   }
-# 	  # }
-# 	  # # truncate to last non-NULL element
-# 	  # changepoints <- changepoints[1:index-1]
-#
-# 	  changepoints <- vector("list", length(aggtypes))
-# 	  for(i in seq_along(aggtypes) ){
-# 	  	log_message(paste0("  By ", aggtypes[i]), showprogress)
-#
-# 	    #    i = 1
-# 	    if( aggtypes[i] == "n"){
-# 	      # TODO: timepoint field should only do is_zero
-# 	      changepoints[[i]] <- list(is_zero=changepoints_is_zero(aggfieldvalues, aggtype = aggtypes[i])
-# 	      )
-# 	    } else if( aggtypes[i] %in% c("missing_n","missing_perc","distinct") ){
-# 	      changepoints[[i]] <- list(is_zero=changepoints_is_zero(aggfieldvalues, aggtype = aggtypes[i], nainterpretation = "ignore"),
-# 	                                is_na=changepoints_is_na(aggfieldvalues, aggtype = aggtypes[i])
-# 	      )
-# 	    } else if( aggtypes[i] %in% c("min","max") ){
-# 	      if( inherits(aggfieldvalues[[aggtypes[i]]], "POSIXct") || inherits(aggfieldvalues[[aggtypes[i]]], "Date") ) {
-# 	        # zeros are not meaningful for dates
[…]
