+++ R/aggregatedata.R
@@ -1,9 +1,6 @@
 # Code for creation of aggregatedata object
 # These contain both the (vector) data for the aggregated field and the relevant metadata
 # TODO: decide whether better to store values as a dataframe or a list of ts/zoo objects
-# NOTE: Changepoints functionality disabled until we find a method that works
-# (have not removed params from internal functions but instead have set them to
-# default to "none")
 # NOTE: Partitionfield functionality disabled until we work out how to present
 # it (creating aligned plots only works if small no. of partitions)
 
@@ -23,7 +20,6 @@ aggregatefield <- function(datafield,
 													 timepointfieldvalues,
 													 alltimepoints,
 													 aggregation_timeunit,
-													 changepointmethods = "none",
 													 partitionfieldname = NULL,
 													 partitionfieldvalue = NULL,
 													 showprogress = TRUE) {
@@ -241,22 +237,12 @@ aggregatefield <- function(datafield,
 		}
 	}
 
-	# TODO: set individual min/max dates per aggfield? Probably shouldn't calculate
-	# changepoints in sections where there are no records for that aggfield NOTE:
-	# Changepoints functionality disabled until we find a method that works
-	# cpts <-
-	# 	find_changepoints(groupedvals,
-	# 										method = changepointmethods,
-	# 										showprogress = showprogress)[[1]]
-
 	log_message(paste0("Finished"), showprogress)
 
 	structure(
 		list(
 			values = groupedvals,
 			functionlist = functionlist,
-			# NOTE: Changepoints functionality disabled until we find a method that works
-			# changepoints = cpts,
 			fieldtype = datafield$fieldtype,
 			columnname = datafield$columnname
 			# NOTE: Partitionfield functionality disabled until we work out how to present it
@@ -282,7 +268,6 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 #'
 #' @param aggfields all aggregatefield objects from data (i.e. excluding
 #'   calculated aggfields)
-#' @param changepointmethods currently disabled
 #' @param partitionfieldname currently disabled
 #' @param partitionfieldvalue currently disabled
 #' @param showprogress Print progress to console
@@ -293,7 +278,6 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 # TODO: this field has a numeric datatype whereas individual fields have an int
 # datatype, decide if need to make them all the same
 aggregateallfields <- function(aggfields,
-															 changepointmethods = "none",
 															 partitionfieldname = NULL,
 															 partitionfieldvalue = NULL,
 															 showprogress = TRUE) {
@@ -337,17 +321,12 @@ aggregateallfields <- function(aggfields,
 	groupedvals[, "missing_perc" := 100 * missing_n / (n + missing_n + nonconformant_n)]
 	groupedvals[, "nonconformant_perc" := 100 * nonconformant_n / (n + missing_n + nonconformant_n)]
 
-	# NOTE: Changepoints functionality disabled until we find a method that works
-#	cpts <- find_changepoints(groupedvals, method = changepointmethods, showprogress = showprogress)[[1]]
-
 	log_message(paste0("Finished"), showprogress)
 
 	structure(
 		list(
 			values = groupedvals,
 			functionlist = functionlist,
-			# NOTE: Changepoints functionality disabled until we find a method that works
[…]
