### 2022-10-05 db31870
Remove unused code - partitionfield functionality. rOpenSci item (r2-22)
