+++ R/aggregatedata.R
@@ -1,14 +1,9 @@
 # Code for creation of aggregatedata object
 # These contain both the (vector) data for the aggregated field and the relevant metadata
 # TODO: decide whether better to store values as a dataframe or a list of ts/zoo objects
-# NOTE: Partitionfield functionality disabled until we work out how to present
-# it (creating aligned plots only works if small no. of partitions)
 
 # -----------------------------------------------------------------------------
 #' Constructor for an individual aggregatefield object
-#'
-#' NOTE: keep datafield and timepointfieldvalues separate as
-#' timepointfieldvalues are updated for subaggregates before being passed in
 #'
 #' @param datafield datafield object
 #' @param timepointfieldvalues all values in timepoint field
@@ -20,8 +15,6 @@ aggregatefield <- function(datafield,
 													 timepointfieldvalues,
 													 alltimepoints,
 													 aggregation_timeunit,
-													 partitionfieldname = NULL,
-													 partitionfieldvalue = NULL,
 													 showprogress = TRUE) {
 
 	# initialise known column names to prevent R CMD check notes
@@ -245,9 +238,6 @@ aggregatefield <- function(datafield,
 			functionlist = functionlist,
 			fieldtype = datafield$fieldtype,
 			columnname = datafield$columnname
-			# NOTE: Partitionfield functionality disabled until we work out how to present it
-			# partitionfieldname = partitionfieldname,
-			# partitionfieldvalue = partitionfieldvalue
 		),
 		class = "aggregatefield"
 	)
@@ -268,8 +258,6 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 #'
 #' @param aggfields all aggregatefield objects from data (i.e. excluding
 #'   calculated aggfields)
-#' @param partitionfieldname currently disabled
-#' @param partitionfieldvalue currently disabled
 #' @param showprogress Print progress to console
 #' @return aggregatefield object, including a data.table where the first column
 #'   is the timepoint group, then one column per aggregationfunction
@@ -278,8 +266,6 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 # TODO: this field has a numeric datatype whereas individual fields have an int
 # datatype, decide if need to make them all the same
 aggregateallfields <- function(aggfields,
-															 partitionfieldname = NULL,
-															 partitionfieldvalue = NULL,
 															 showprogress = TRUE) {
 
 	# initialise known column names to prevent R CMD check notes
@@ -329,9 +315,6 @@ aggregateallfields <- function(aggfields,
 			functionlist = functionlist,
 			fieldtype = ft,
 			columnname = "[ALLFIELDSCOMBINED]"
-			# NOTE: Partitionfield functionality disabled until we work out how to present it
-			# partitionfieldname = partitionfieldname,
-			# partitionfieldvalue = partitionfieldvalue
 		),
 		class = "aggregatefield"
 	)
@@ -449,7 +432,6 @@ aggregate_data <- function(sourcedata,
 	agg[[sourcedata$cols_imported_n + 2]] <-
 		aggregateallfields(
 			agg[1:sourcedata$cols_imported_n],
-			aggregation_timeunit,
 			showprogress = showprogress
 		)
 	names(agg) <-
[…]
