### 2022-10-05 d986e84
Remove unused code - option to use fread as parser. rOpenSci item (r2-22)
