+++ R/main.R
@@ -274,44 +274,3 @@ read_data <- function(file,
 	)
 
 }
-
-# Alternative read_data using fread instead
-# #' Read data from delimited file (data.table)
-# #'
-# #' Read rectangular data from a delimited file, with all columns read in as character type so that the package can later detect any non-conformant values.
-# #' Operates as a restricted implementation of \code{data.table::fread}, but returns a data.frame by default.
-# #'
-# #' @param file A string containing path of file containing data to load,
-# #' or a URL starting http://, file://, etc. Compressed files with extension ‘.gz’ and ‘.bz2’ are supported.
-# #' @param sep The separator between columns. Defaults to the character in the set
-# #' [,\t |;:] that separates the sample of rows into the most number of lines with the same number of fields
-# #' @param header Does the first row contain the column names?
-# #' @param skip Header rows to skip
-# #' @param nrows	The maximum number of rows to read
-# #' @param select A vector of column names or numbers to keep, drop the rest
-# #' @param drop A vector of column names or numbers to drop, keep the rest
-# #' @param quote Single character used to quote strings
-# #' @param strip.white Strip leading and trailing whitespaces of unquoted fields. Default = TRUE. If FALSE, only header trailing spaces are removed.
-# #' @param showprogress Print progress to console. Default = TRUE
-# #' @return A data.frame by default, or a \code{data.table} object if \code{data.table} = TRUE
-# #' @examples sourcedataobj <- read_data(
-# #'   system.file("extdata", "example_data.csv", package = "daiquiri"),
-# #'   sep = ","
-# #'   header = TRUE
-# #' )
-# #' @seealso \code{\link{fieldtypes}}, \code{\link{availablefieldtypes}}, \code{\link{aggregate_data}}, \code{\link{report_data}}, \code{\link{create_report}}
-# #' @export
-# fread_data <- function(file, sep = "auto", header = "auto", select = NULL, drop = NULL,
-# 												 skip = 0, nrows = Inf, quote = "\"",
-# 												 strip.white = TRUE, data.table = FALSE, showprogress = TRUE){
-# 	#	NOTE: col.names param has no default for consistency with read.table and fread
-# 	# 	call_args = names(match.call())
-# 	#	(!'col.names' %chin% call_args)
-# 		data.table::fread(file, sep = sep, nrows = nrows, header = header, skip = skip, select = select,
-# 																	 drop = drop, check.names = FALSE,
-# 																	 quote = quote, strip.white = strip.white,
-# 																	 blank.lines.skip = TRUE, colClasses = "character",
-# 											data.table = data.table, showProgress = showprogress)
-# 	}
-#
-#
