### 2022-10-05 e4f5b4e
Remove unused code - misc. rOpenSci item (r2-22)
