+++ tests/testthat/test-utilities.R
@@ -2,10 +2,6 @@ test_that("validate_params_required() checks if params are supplied", {
 	# NOTE: testfn_params_required() defined in utilities.R as devtools::test() can't find it when it's defined here
 	# 	Think it is something to do with environments but haven't figured it out
 
-	# testfn_params_required <- function(p1, p2, p3 = NULL) {
-	# 	validate_params_required(match.call())
-	# }
-
 	expect_silent(testfn_params_required(1, 2))
 	expect_silent(testfn_params_required(p2 = 1, p1 = 2))
 
