### 2022-10-10 7fd3143
Remove unnecessary literal uses of "NOTE" in vignette. rOpenSci item (r02-01)
