### 2022-10-10 0858915
Rename the example dataset from example_data to example_prescriptions. rOpenSci item (r2-21)
