+++ R/main.R
@@ -69,7 +69,7 @@ NULL
 #' \donttest{
 #' # load example data into a data.frame
 #' rawdata <- read_data(
-#'   system.file("extdata", "example_data.csv", package = "daiquiri"),
+#'   system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
 #'   delim = ",",
 #'   col_names = TRUE
 #' )
@@ -234,7 +234,7 @@ print.daiquiri_object <- function(x, ...) {
 #' @param showprogress Display a progress bar? Default = TRUE
 #' @return A data frame
 #' @examples rawdata <- read_data(
-#'   system.file("extdata", "example_data.csv", package = "daiquiri"),
+#'   system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
 #'   delim = ",",
 #'   col_names = TRUE
 #' )
