+++ R/sourcedata.R
@@ -53,7 +53,7 @@ is.datafield <- function(x) inherits(x, "datafield")
 #' @examples
 #' # load example data into a data.frame
 #' rawdata <- read_data(
-#'   system.file("extdata", "example_data.csv", package = "daiquiri"),
+#'   system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
 #'   delim = ",",
 #'   col_names = TRUE
 #' )
