+++ R/reports.R
@@ -21,7 +21,7 @@
 #' \donttest{
 #' # load example data into a data.frame
 #' rawdata <- read_data(
-#'   system.file("extdata", "example_data.csv", package = "daiquiri"),
+#'   system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
 #'   delim = ",",
 #'   col_names = TRUE
 #' )
