+++ man/prepare_data.Rd
@@ -45,7 +45,7 @@ prepare for aggregation.
 \examples{
 # load example data into a data.frame
 rawdata <- read_data(
-  system.file("extdata", "example_data.csv", package = "daiquiri"),
+  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
   delim = ",",
   col_names = TRUE
 )
