+++ man/report_data.Rd
@@ -41,7 +41,7 @@ Generate report from previously-created sourcedata and aggregatedata objects
 \donttest{
 # load example data into a data.frame
 rawdata <- read_data(
-  system.file("extdata", "example_data.csv", package = "daiquiri"),
+  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
   delim = ",",
   col_names = TRUE
 )
