+++ man/aggregate_data.Rd
@@ -26,7 +26,7 @@ Default time period for aggregation is a calendar day
 \examples{
 # load example data into a data.frame
 rawdata <- read_data(
-  system.file("extdata", "example_data.csv", package = "daiquiri"),
+  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
   delim = ",",
   col_names = TRUE
 )
