+++ man/export_aggregated_data.Rd
@@ -27,7 +27,7 @@ field in dataset.
 }
 \examples{
 rawdata <- read_data(
-  system.file("extdata", "example_data.csv", package = "daiquiri"),
+  system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
   delim = ",",
   col_names = TRUE
 )

[…]
