### 2022-10-11 f05f70b
close_log() now returns the path to the closed log file (if any)
