+++ man/close_log.Rd
@@ -6,6 +6,10 @@
 \usage{
 close_log()
 }
+\value{
+If a log file was found, the path to the log file that was closed,
+  otherwise an empty string
+}
 \description{
 Closes any active log file
 }
