### 2022-10-11 b66ca9a
Document all return values and update functions to return invisibly where appropriate. rOpenSci item (r2-02)
