### 2022-10-11 7f52bca
Move package documentation to R/daiquiri-package.R using usethis::use_package_doc(). rOpenSci item (r2-04)
