+++ man/daiquiri-package.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/daiquiri-package.R
+\docType{package}
+\name{daiquiri-package}
+\alias{daiquiri}
+\alias{daiquiri-package}
+\title{daiquiri: Data Quality Reporting for Temporal Datasets}
+\description{
+Generate reports that enable quick visual review of temporal shifts in record-level data. Time series plots showing aggregated values are automatically created for each data field (column) depending on its contents (e.g. min/max/mean values for numeric data, no. of distinct values for categorical data), as well as overviews for missing values, non-conformant values, and duplicated rows. The resulting reports are sharable and can contribute to forming a transparent record of the entire analysis process. It is designed with Electronic Health Records in mind, but can be used for any type of record-level temporal data (i.e. tabular data where each row represents a single “event”, one column contains the "event date", and other columns contain any associated values for the event).
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://github.com/phuongquan/daiquiri}
+  \item Report bugs at \url{https://github.com/phuongquan/daiquiri/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: T. Phuong Quan \email{phuong.quan@ndm.ox.ac.uk} (\href{https://orcid.org/0000-0001-8566-1817}{ORCID})
+
+Other contributors:
+\itemize{
+  \item Jack Cregan [contributor]
+  \item  University of Oxford [copyright holder]
+  \item  National Institute for Health Research (NIHR) [funder]
+}
+
+}
+\keyword{internal}
