### 2022-10-11 0526c90
Move package documentation to R/daiquiri-package.R using usethis::use_package_doc() (cont.). rOpenSci item (r2-04)
