### 2022-10-12 ac191e1
Enable markdown features in Roxygen2 and update documentation to format code and links where missing. rOpenSci item (r2-03, r2-05)
