+++ R/fieldtypes.R
@@ -84,11 +84,12 @@ is.fieldtype_calculated <- function(x) inherits(x, c("fieldtype_allfields", "fie
 
 #' Available fieldtypes
 #'
-#' Each column in the source dataset must be assigned to one of these
-#' fieldtypes, through a \code{\link{fieldtypes}} specification.
-#' @seealso \code{\link{fieldtypes}}
+#' Each column in the source dataset must be assigned to a particular `ft_xx`
+#' depending on the type of data that it contains. This is done through a
+#' [fieldtypes()] specification.
+#' @seealso [fieldtypes()]
 #' @name availablefieldtypes
-#' @return A \code{fieldtype} object denoting the type of data in the column
+#' @return A `fieldtype` object denoting the type of data in the column
 #' @examples fts <- fieldtypes(PatientID = ft_uniqueidentifier(),
 #'                             TestID = ft_ignore(),
 #'                             TestDate = ft_timepoint(),
@@ -99,17 +100,17 @@ is.fieldtype_calculated <- function(x) inherits(x, c("fieldtype_allfields", "fie
 #'                             Location = ft_categorical())
 NULL
 
-#' @section Details: \code{ft_timepoint} - identifies the data field which
+#' @section Details: `ft_timepoint()` - identifies the data field which
 #'   should be used as the independent time variable. There should be one and
 #'   only one of these specified.
-#' @param includes_time If TRUE, additional aggregated values will be generated
+#' @param includes_time If `TRUE`, additional aggregated values will be generated
 #'   using the time portion (and if no time portion is present then midnight
-#'   will be assumed). If FALSE, aggregated values will ignore any time portion.
-#'   Default = TRUE
+#'   will be assumed). If `FALSE`, aggregated values will ignore any time portion.
+#'   Default = `TRUE`
 #' @param format Where datetime values are not in the format `YYYY-MM-DD` or
 #'   `YYYY-MM-DD HH:MM:SS`, an alternative format can be specified at the per
-#'   field level, using readr's \code{\link[readr]{col_datetime}} format
-#'   specifications, e.g. \code{format = "\%d/\%m/\%Y"}. When a format is
+#'   field level, using [readr::col_datetime()] format
+#'   specifications, e.g. `format = "%d/%m/%Y"`. When a format is
 #'   supplied, it must match the complete string.
 #' @rdname availablefieldtypes
 #' @export
@@ -132,7 +133,7 @@ ft_timepoint <- function(includes_time = TRUE,
 	)
 }
 
-#' @section Details: \code{ft_uniqueidentifier} - identifies data fields which
+#' @section Details: `ft_uniqueidentifier()` - identifies data fields which
 #'   contain a (usually computer-generated) identifier for an entity, e.g. a
 #'   patient. It does not need to be unique within the dataset.
 #' @rdname availablefieldtypes
@@ -152,12 +153,12 @@ ft_uniqueidentifier <- function() {
 
 }
 
-#' @section Details: \code{ft_categorical} - identifies data fields which should
+#' @section Details: `ft_categorical()` - identifies data fields which should
 #'   be treated as categorical.
-#' @param aggregate_by_each_category If TRUE, aggregated values will be
+#' @param aggregate_by_each_category If `TRUE`, aggregated values will be
 #'   generated for each distinct subcategory as well as for the field overall.
-#'   If FALSE, aggregated values will only be generated for the field overall.
-#'   Default = FALSE
+#'   If `FALSE`, aggregated values will only be generated for the field overall.
+#'   Default = `FALSE`
 #' @rdname availablefieldtypes
 #' @export
 ft_categorical <- function(aggregate_by_each_category = FALSE) {
[…]
