+++ man/report_data.Rd
@@ -15,27 +15,27 @@ report_data(
 }
 \arguments{
 \item{sourcedata}{A \code{sourcedata} object returned from
-\code{\link{prepare_data}} function}
+\code{\link[=prepare_data]{prepare_data()}} function}
 
 \item{aggregatedata}{An \code{aggregatedata} object returned from
-\code{\link{aggregate_data}} function}
+\code{\link[=aggregate_data]{aggregate_data()}} function}
 
 \item{save_directory}{String specifying directory in which to save the
 report. Default is current directory.}
 
 \item{save_filename}{String specifying filename for the report, excluding any
 file extension. If no filename is supplied, one will be automatically
-generated with the format daiquiri_report_YYMMDD_HHMMSS.}
+generated with the format \code{daiquiri_report_YYMMDD_HHMMSS}.}
 
-\item{format}{File format of the report. Currently only "html" is supported}
+\item{format}{File format of the report. Currently only \code{"html"} is supported}
 
-\item{showprogress}{Print progress to console. Default = TRUE}
+\item{showprogress}{Print progress to console. Default = \code{TRUE}}
 }
 \value{
 A string containing the name and path of the saved report
 }
 \description{
-Generate report from previously-created sourcedata and aggregatedata objects
+Generate report from previously-created \code{sourcedata} and \code{aggregatedata} objects
 }
 \examples{
 \donttest{
@@ -84,6 +84,6 @@ report_data(
 
 }
 \seealso{
-\code{\link{prepare_data}}, \code{\link{aggregate_data}},
-  \code{\link{create_report}}
+\code{\link[=prepare_data]{prepare_data()}}, \code{\link[=aggregate_data]{aggregate_data()}},
+\code{\link[=create_report]{create_report()}}
 }
