### 2022-10-12 d222d1a
Remove messages from close_log(), just set the return value instead
