### 2022-10-12 e3b1f82
Split tests that tested multiple behaviours into smaller units. rOpenSci item (r2-10, r2-12)
