+++ tests/testthat/test-utilities.R
@@ -1,9 +1,13 @@
-test_that("validate_params_required() checks if params are supplied", {
+test_that("validate_params_required() checks is silent if required params are supplied", {
 	# NOTE: testfn_params_required() defined in utilities.R as devtools::test() can't find it when it's defined here
 	# 	Think it is something to do with environments but haven't figured it out
 
 	expect_silent(testfn_params_required(1, 2))
 	expect_silent(testfn_params_required(p2 = 1, p1 = 2))
+})
+
+test_that("validate_params_required() checks returns an error if any required params are not supplied", {
+	# NOTE: testfn_params_required() defined in utilities.R as devtools::test() can't find it when it's defined here
 
 	expect_error(testfn_params_required(),
 							 class = "invalid_param_missing")
@@ -15,34 +19,42 @@ test_that("validate_params_required() checks if params are supplied", {
 })
 
 
-test_that("validate_params_type() checks if params are of correct type", {
+test_that("validate_params_type() is silent if all params are of correct type", {
 	# NOTE: testfn_params_type() defined in utilities.R as devtools::test() can't find it when it's defined here
-	# 	Think it is something to do with environments but haven't figured it out
-	# TODO: consider splitting this into multiple tests
 
 	# all default args are valid
 	expect_silent(testfn_params_type())
+})
 
+test_that("validate_params_type() checks df params are of correct type", {
 	expect_error(testfn_params_type(df = c("Fieldname" = 123)),
 							 class = "invalid_param_type")
+})
 
+test_that("validate_params_type() checks override_columnnames params are of correct type", {
 	expect_error(testfn_params_type(override_columnnames = c("col1", "col2")),
 							 class = "invalid_param_type")
+})
 
+test_that("validate_params_type() checks na params are of correct type", {
 	expect_error(testfn_params_type(na = NA),
 							 class = "invalid_param_type")
 	expect_error(testfn_params_type(na = c(NULL)),
 							 class = "invalid_param_type")
 	expect_error(testfn_params_type(na = c(1, 2, 3)),
 							 class = "invalid_param_type")
+})
 
+test_that("validate_params_type() checks dataset_shortdesc params are of correct type", {
 	expect_silent(testfn_params_type(dataset_shortdesc = ""))
 	expect_silent(testfn_params_type(dataset_shortdesc = NULL))
 	expect_error(testfn_params_type(dataset_shortdesc = 123),
 							 class = "invalid_param_type")
 	expect_error(testfn_params_type(dataset_shortdesc = c("col1", "col2")),
 							 class = "invalid_param_type")
+})
 
+test_that("validate_params_type() checks aggregation_timeunit params are one of day/week/month/quarter/year", {
 	expect_silent(testfn_params_type(aggregation_timeunit = "day"))
 	expect_silent(testfn_params_type(aggregation_timeunit = "week"))
 	expect_silent(testfn_params_type(aggregation_timeunit = "month"))
@@ -50,9 +62,14 @@ test_that("validate_params_type() checks if params are of correct type", {
 	expect_silent(testfn_params_type(aggregation_timeunit = "year"))
 	expect_error(testfn_params_type(aggregation_timeunit = "other"),
 							 class = "invalid_param_type")
+})
+
+test_that("validate_params_type() checks aggregation_timeunit params cannot be a vector", {
 	expect_error(testfn_params_type(aggregation_timeunit = c("day", "week")),
 							 class = "invalid_param_type")
+})
 
+test_that("validate_params_type() checks save_directory params are of correct type", {
 	# Real dir
 	expect_silent(testfn_params_type(save_directory = test_path()))
 	# Real dir with trailing slash
[…]

[…]
