### 2022-10-13 579fd90
Remove unused code - misc. rOpenSci item (r2-22)
