### 2022-10-13 4aeb7a9
Run styler on whole package. rOpenSci item (r2-15)
Had to temporarily remove \dontshow{file.remove("./example_data_report.html")} from main.R and reports.R examples as styler raised a warning and did not update them with that line in.
