+++ R/reports.R
@@ -29,16 +29,18 @@
 #' # validate and prepare the data for aggregation
 #' sourcedataobj <- prepare_data(
 #'   rawdata,
-#'   fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
+#'   fieldtypes = fieldtypes(
+#'     PrescriptionID = ft_uniqueidentifier(),
 #'     PrescriptionDate = ft_timepoint(),
 #'     AdmissionDate = ft_datetime(includes_time = FALSE),
 #'     Drug = ft_freetext(),
 #'     Dose = ft_numeric(),
 #'     DoseUnit = ft_categorical(),
 #'     PatientID = ft_ignore(),
-#'     Location = ft_categorical(aggregate_by_each_category=TRUE)),
+#'     Location = ft_categorical(aggregate_by_each_category = TRUE)
+#'   ),
 #'   override_columnnames = FALSE,
-#'   na = c("","NULL"),
+#'   na = c("", "NULL"),
 #'   dataset_shortdesc = "Example data provided with package",
 #'   showprogress = TRUE
 #' )
@@ -66,57 +68,56 @@
 #'   [create_report()]
 #' @export
 report_data <- function(sourcedata,
-												aggregatedata,
-												save_directory = ".",
-												save_filename = NULL,
-												format = "html",
-												showprogress = TRUE) {
+                        aggregatedata,
+                        save_directory = ".",
+                        save_filename = NULL,
+                        format = "html",
+                        showprogress = TRUE) {
+  log_function_start(match.call()[[1]])
 
-	log_function_start(match.call()[[1]])
+  validate_params_required(match.call())
+  validate_params_type(match.call(),
+    sourcedata = sourcedata,
+    aggregatedata = aggregatedata,
+    save_directory = save_directory,
+    save_filename = save_filename,
+    showprogress = showprogress,
+    format = format
+  )
 
-	validate_params_required(match.call())
-	validate_params_type(match.call(),
-											 sourcedata = sourcedata,
-											 aggregatedata = aggregatedata,
-											 save_directory = save_directory,
-											 save_filename = save_filename,
-											 showprogress = showprogress,
-											 format = format)
+  if (is.null(save_filename)) {
+    save_filename <-
+      paste0("daiquiri_report_", format(Sys.time(), "%Y%m%d%_%H%M%S"))
+  }
 
-	if (is.null(save_filename)) {
-		save_filename <-
-			paste0("daiquiri_report_", format(Sys.time(), "%Y%m%d%_%H%M%S"))
-	}
+  fileandpath <- file.path(save_directory, paste0(save_filename, ".html"))
 
-	fileandpath <- file.path(save_directory, paste0(save_filename, ".html"))
+  if (format == "html") {
+    log_message("Generating html report...", showprogress)
+    rmarkdown::render(
+      input = system.file(
+        "rmd",
+        "report_htmldoc.Rmd",
+        package = utils::packageName(),
+        mustWork = TRUE
+      ),
+      output_file = paste0(save_filename, ".html"),
+      output_dir = save_directory,
+      params = list(sourcedata = sourcedata, aggregatedata = aggregatedata),
+      quiet = !showprogress
+    )
+  } else {
+    stop(paste(
+      "Invalid format: ", format,
+      ". Only html format is currently supported"
+    ))
+  }
 
-	if (format == "html") {
-		log_message("Generating html report...", showprogress)
-		rmarkdown::render(
-			input = system.file(
-								"rmd",
[…]
