+++ R/main.R
@@ -50,16 +50,18 @@
 #' # create a report in the current directory
 #' daiqobj <- create_report(
 #'   rawdata,
-#'   fieldtypes = fieldtypes(PrescriptionID = ft_uniqueidentifier(),
+#'   fieldtypes = fieldtypes(
+#'     PrescriptionID = ft_uniqueidentifier(),
 #'     PrescriptionDate = ft_timepoint(),
 #'     AdmissionDate = ft_datetime(includes_time = FALSE),
 #'     Drug = ft_freetext(),
 #'     Dose = ft_numeric(),
 #'     DoseUnit = ft_categorical(),
 #'     PatientID = ft_ignore(),
-#'     Location = ft_categorical(aggregate_by_each_category=TRUE)),
+#'     Location = ft_categorical(aggregate_by_each_category = TRUE)
+#'   ),
 #'   override_columnnames = FALSE,
-#'   na = c("","NULL"),
+#'   na = c("", "NULL"),
 #'   dataset_shortdesc = "Example data provided with package",
 #'   aggregation_timeunit = "day",
 #'   save_directory = ".",
@@ -75,109 +77,114 @@
 #'   [availablefieldtypes()]
 #' @export
 create_report <- function(df,
-													fieldtypes,
-													override_columnnames = FALSE,
-													na = c("", "NA", "NULL"),
-													dataset_shortdesc = NULL,
-													aggregation_timeunit = "day",
-													save_directory = ".",
-													save_filename = NULL,
-													showprogress = TRUE,
-													log_directory = NULL) {
+                          fieldtypes,
+                          override_columnnames = FALSE,
+                          na = c("", "NA", "NULL"),
+                          dataset_shortdesc = NULL,
+                          aggregation_timeunit = "day",
+                          save_directory = ".",
+                          save_filename = NULL,
+                          showprogress = TRUE,
+                          log_directory = NULL) {
 
-	# if a log directory is supplied, start a new log. Otherwise, close any existing log
-	if (!is.null(log_directory)) {
-		log_filename <- initialise_log(log_directory)
-	} else{
-		close_log()
-		log_filename <- NULL
-	}
+  # if a log directory is supplied, start a new log. Otherwise, close any existing log
+  if (!is.null(log_directory)) {
+    log_filename <- initialise_log(log_directory)
+  } else {
+    close_log()
+    log_filename <- NULL
+  }
 
-	log_function_start(match.call()[[1]])
+  log_function_start(match.call()[[1]])
 
-	# check params before running anything so that it fails sooner rather than later
-	validate_params_required(match.call())
-	validate_params_type(
-		match.call(),
-		df = df,
-		fieldtypes = fieldtypes,
-		override_columnnames = override_columnnames,
-		na = na,
-		dataset_shortdesc = dataset_shortdesc,
-		aggregation_timeunit = aggregation_timeunit,
-		save_directory = save_directory,
-		save_filename = save_filename,
-		showprogress = showprogress,
-		log_directory = log_directory
-	)
+  # check params before running anything so that it fails sooner rather than later
+  validate_params_required(match.call())
+  validate_params_type(
+    match.call(),
+    df = df,
+    fieldtypes = fieldtypes,
+    override_columnnames = override_columnnames,
+    na = na,
+    dataset_shortdesc = dataset_shortdesc,
+    aggregation_timeunit = aggregation_timeunit,
[…]

[…]
