### 2022-10-13 300ec04
Rename fieldtype/fieldtypes as field_type/field_types. rOpenSci item (r2-19)
