+++ R/fieldtypes.R
@@ -1,6 +1,6 @@
 # Create field specification (don't know whether to call them fields or columns)
 #
-# NOTE: ideally want classes to inherit, e.g. all fieldtypes (except "ignored"
+# NOTE: ideally want classes to inherit, e.g. all field_types (except "ignored"
 # ones should implement a function for checking missing data)
 #
 # copying structure used in readr
@@ -8,21 +8,21 @@
 # -----------------------------------------------------------------------------
 # SPECIFY ALLOWABLE TYPES
 
-#' Constructor for individual fieldtype object
+#' Constructor for individual field_type object
 #'
-#' @param type string denoting the fieldtype
+#' @param type string denoting the field_type
 #' @param collector readr `collector' to use when parsing the data
 #' @param dataclass type of data, e.g. character or POSIXct
 #' @param aggfunctions aggfunctions to apply to the data in this field
 #' @param options additional options for certain aggfunctions
-#' @return fieldtype object
+#' @return field_type object
 #' @noRd
 # TODO: decide whether to require all aggfunctions to be supplied or automatically include the basic ones
-fieldtype <- function(type,
-                      collector,
-                      dataclass,
-                      aggfunctions = c("n", "missing_n", "missing_perc"),
-                      options = NULL) {
+field_type <- function(type,
+                       collector,
+                       dataclass,
+                       aggfunctions = c("n", "missing_n", "missing_perc"),
+                       options = NULL) {
   structure(
     list(
       type = type,
@@ -31,62 +31,62 @@ fieldtype <- function(type,
       aggfunctions = aggfunctions,
       options = options
     ),
-    class = c(paste0("fieldtype_", type), "fieldtype")
+    class = c(paste0("field_type_", type), "field_type")
   )
 }
 
-#' Test if object is a fieldtype object
+#' Test if object is a field_type object
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
-is.fieldtype <- function(x) inherits(x, "fieldtype")
+is.field_type <- function(x) inherits(x, "field_type")
 
-#' Test if object is a timepoint fieldtype
+#' Test if object is a timepoint field_type
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
-is.fieldtype_timepoint <- function(x) inherits(x, "fieldtype_timepoint")
+is.field_type_timepoint <- function(x) inherits(x, "field_type_timepoint")
 
-#' Test if object is a ignore fieldtype
+#' Test if object is a ignore field_type
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
-is.fieldtype_ignore <- function(x) inherits(x, "fieldtype_ignore")
+is.field_type_ignore <- function(x) inherits(x, "field_type_ignore")
 
-#' Test if object is a datetime fieldtype
+#' Test if object is a datetime field_type
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
-is.fieldtype_datetime <- function(x) inherits(x, "fieldtype_datetime")
+is.field_type_datetime <- function(x) inherits(x, "field_type_datetime")
 
-#' Test if object is a numeric fieldtype
+#' Test if object is a numeric field_type
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
-is.fieldtype_numeric <- function(x) inherits(x, "fieldtype_numeric")
+is.field_type_numeric <- function(x) inherits(x, "field_type_numeric")
 
-#' Test if object is a calculated fieldtype
+#' Test if object is a calculated field_type
 #'
 #' @param x object to test
 #' @return Logical
 #' @noRd
[…]
