+++ R/aggregatedata.R
@@ -421,7 +421,7 @@ aggregateallfields <- function(aggfields,
 #'     PatientID = ft_ignore(),
 #'     Location = ft_categorical(aggregate_by_each_category = TRUE)
 #'   ),
-#'   override_columnnames = FALSE,
+#'   override_column_names = FALSE,
 #'   na = c("", "NULL")
 #' )
 #'
@@ -568,7 +568,7 @@ is.aggregatedata <- function(x) inherits(x, "aggregatedata")
 #'     PatientID = ft_ignore(),
 #'     Location = ft_categorical(aggregate_by_each_category = TRUE)
 #'   ),
-#'   override_columnnames = FALSE,
+#'   override_column_names = FALSE,
 #'   na = c("", "NULL")
 #' )
 #'
