### 2022-10-13 628b19f
Rename showprogress to show_progress. rOpenSci item (r2-15)
