+++ R/aggregatedata.R
@@ -15,17 +15,17 @@ aggregatefield <- function(datafield,
                            timepointfieldvalues,
                            alltimepoints,
                            aggregation_timeunit,
-                           showprogress = TRUE) {
+                           show_progress = TRUE) {
 
   # initialise known column names to prevent R CMD check notes
   n <- value <- values <- timepointgroup <- NULL
 
-  log_message(paste0("Preparing..."), showprogress)
+  log_message(paste0("Preparing..."), show_progress)
   functionlist <- datafield$field_type$aggfunctions
 
   log_message(
     paste0("Aggregating ", get_datafield_basetype(datafield), " field..."),
-    showprogress
+    show_progress
   )
 
   # TODO: consider doing this by reference
@@ -44,7 +44,7 @@ aggregatefield <- function(datafield,
 
   for (i in seq_along(functionlist)) {
     f <- functionlist[i]
-    log_message(paste0("  By ", f), showprogress)
+    log_message(paste0("  By ", f), show_progress)
     if (f == "n") {
       # number of values present (including non-conformant ones)
       groupedvals[datafield_dt[, list("value" = sum(!(is.na(values) &
@@ -125,12 +125,12 @@ aggregatefield <- function(datafield,
       # create a separate column per category value
       distinctcategories <-
         sort(datafield_dt[is.na(values) == FALSE, unique(values)])
-      log_message(paste0("    ", length(distinctcategories), " categories found"), showprogress)
+      log_message(paste0("    ", length(distinctcategories), " categories found"), show_progress)
       # If there is only one category, don't bother
       if (length(distinctcategories) > 1) {
         # TODO: consider setting a max number of categories
         for (j in seq_along(distinctcategories)) {
-          log_message(paste0("    ", j, ": ", distinctcategories[j]), showprogress)
+          log_message(paste0("    ", j, ": ", distinctcategories[j]), show_progress)
           catval <- distinctcategories[j]
           catname <-
             paste0(f, "_", j, "_", gsub("([[:punct:]])|\\s+", "_", catval))
@@ -295,7 +295,7 @@ aggregatefield <- function(datafield,
     }
   }
 
-  log_message(paste0("Finished"), showprogress)
+  log_message(paste0("Finished"), show_progress)
 
   structure(
     list(
@@ -323,7 +323,7 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 #'
 #' @param aggfields all aggregatefield objects from data (i.e. excluding
 #'   calculated aggfields)
-#' @param showprogress Print progress to console
+#' @param show_progress Print progress to console
 #' @return aggregatefield object, including a data.table where the first column
 #'   is the timepoint group, then one column per aggregationfunction
 #' @noRd
@@ -331,7 +331,7 @@ is.aggregatefield <- function(x) inherits(x, "aggregatefield")
 # TODO: this field has a numeric datatype whereas individual fields have an int
 # datatype, decide if need to make them all the same
 aggregateallfields <- function(aggfields,
-                               showprogress = TRUE) {
+                               show_progress = TRUE) {
 
   # initialise known column names to prevent R CMD check notes
   n <- missing_n <- nonconformant_n <- NULL
@@ -374,7 +374,7 @@ aggregateallfields <- function(aggfields,
   groupedvals[, "missing_perc" := 100 * missing_n / (n + missing_n + nonconformant_n)]
   groupedvals[, "nonconformant_perc" := 100 * nonconformant_n / (n + missing_n + nonconformant_n)]
 
[…]
