+++ R/fieldtypes.R
@@ -12,23 +12,23 @@
 #'
 #' @param type string denoting the field_type
 #' @param collector readr `collector' to use when parsing the data
-#' @param dataclass type of data, e.g. character or POSIXct
-#' @param aggfunctions aggfunctions to apply to the data in this field
-#' @param options additional options for certain aggfunctions
+#' @param data_class type of data, e.g. character or POSIXct
+#' @param aggregation_functions aggregation_functions to apply to the data in this field
+#' @param options additional options for certain aggregation_functions
 #' @return field_type object
 #' @noRd
-# TODO: decide whether to require all aggfunctions to be supplied or automatically include the basic ones
+# TODO: decide whether to require all aggregation_functions to be supplied or automatically include the basic ones
 field_type <- function(type,
                        collector,
-                       dataclass,
-                       aggfunctions = c("n", "missing_n", "missing_perc"),
+                       data_class,
+                       aggregation_functions = c("n", "missing_n", "missing_perc"),
                        options = NULL) {
   structure(
     list(
       type = type,
       collector = collector,
-      dataclass = dataclass,
-      aggfunctions = aggfunctions,
+      data_class = data_class,
+      aggregation_functions = aggregation_functions,
       options = options
     ),
     class = c(paste0("field_type_", type), "field_type")
@@ -115,18 +115,18 @@ NULL
 ft_timepoint <- function(includes_time = TRUE,
                          format = "") {
   # NOTE: nonconformant values appear in validation warnings
-  aggfn <- c("n")
+  agg_fun <- c("n")
   options <- NULL
   if (includes_time) {
-    aggfn <- c(aggfn, "midnight_n", "midnight_perc")
+    agg_fun <- c(agg_fun, "midnight_n", "midnight_perc")
     # TODO: can probably do something more sophisticated here with match.call()
     options <- "includes_time"
   }
   field_type(
     type = "timepoint",
     collector = readr::col_datetime(format = format),
-    dataclass = "POSIXct",
-    aggfunctions = aggfn,
+    data_class = "POSIXct",
+    aggregation_functions = agg_fun,
     options = options
   )
 }
@@ -137,18 +137,18 @@ ft_timepoint <- function(includes_time = TRUE,
 #' @rdname field_types_available
 #' @export
 ft_uniqueidentifier <- function() {
-  # TODO: potential additional aggfunctions: proportion numeric; distinct first chars
+  # TODO: potential additional aggregation_functions: proportion numeric; distinct first chars
   field_type(
     type = "uniqueidentifier",
     collector = readr::col_character(),
-    dataclass = "character",
-    aggfunctions = c(
+    data_class = "character",
+    aggregation_functions = c(
       "n",
       "missing_n",
       "missing_perc",
-      "minlength",
-      "maxlength",
-      "meanlength"
+      "min_length",
+      "max_length",
+      "mean_length"
     )
   )
 }
@@ -163,18 +163,18 @@ ft_uniqueidentifier <- function() {
 #' @export
 ft_categorical <- function(aggregate_by_each_category = FALSE) {
   # TODO: allow more options for aggregate_by_each_category, e.g. topx (bysize), or accept a vector of values
-  aggfn <- c("n", "missing_n", "missing_perc", "distinct")
+  agg_fun <- c("n", "missing_n", "missing_perc", "distinct")
   options <- NULL
   if (aggregate_by_each_category) {
-    aggfn <- c(aggfn, "subcat_n", "subcat_perc")
[…]
