+++ man/export_aggregated_data.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aggregatedata.R
+% Please edit documentation in R/aggregate_data.R
 \name{export_aggregated_data}
 \alias{export_aggregated_data}
 \title{Export aggregated data}
