+++ man/print_field_types_template.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/fieldtypes.R
+% Please edit documentation in R/field_types.R
 \name{print_field_types_template}
 \alias{print_field_types_template}
 \title{Print a template field_types() specification to console}
