+++ man/field_types_available.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/fieldtypes.R
+% Please edit documentation in R/field_types.R
 \name{field_types_available}
 \alias{field_types_available}
 \alias{ft_timepoint}
