+++ man/field_types.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/fieldtypes.R
+% Please edit documentation in R/field_types.R
 \name{field_types}
 \alias{field_types}
 \title{Create field_types specification}
