+++ man/prepare_data.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sourcedata.R
+% Please edit documentation in R/source_data.R
 \name{prepare_data}
 \alias{prepare_data}
 \title{Prepare source data}
