### 2022-10-14 c3d5fb7
Print the result of examples. rOpenSci item (r2-07)
