### 2022-10-14 2d2d988
Reviewed custom classes. rOpenSci item (r2-29)
