### 2022-10-14 d7511a1
Add rOpenSci reviewers and documentation site url.
