### 2022-10-14 cb4c703
Reorganise functions within files. rOpenSci item (r2-15)
