+++ R/aggregate_data.R
@@ -1,6 +1,288 @@
 # Code for creation of aggregated_data object
 # These contain both the (vector) data for the aggregated field and the relevant metadata
-# TODO: decide whether better to store values as a dataframe or a list of ts/zoo objects
+
+# -----------------------------------------------------------------------------
+#' Aggregate source data
+#'
+#' Aggregates a `daiquiri_source_data` object based on the [field_types()] specified at load time.
+#' Default time period for aggregation is a calendar day
+#'
+#' @param source_data A `daiquiri_source_data` object returned from
+#'   [prepare_data()] function
+#' @param aggregation_timeunit Unit of time to aggregate over. Specify one of
+#'   `"day"`, `"week"`, `"month"`, `"quarter"`, `"year"`. The `"week"` option is
+#'   Monday-based. Default = `"day"`
+#' @param show_progress Print progress to console. Default = `TRUE`
+#' @return A `daiquiri_aggregated_data` object
+#' @examples
+#' # load example data into a data.frame
+#' raw_data <- read_data(
+#'   system.file("extdata", "example_prescriptions.csv", package = "daiquiri"),
+#'   delim = ",",
+#'   col_names = TRUE
+#' )
+#'
+#' # validate and prepare the data for aggregation
+#' source_data <- prepare_data(
+#'   raw_data,
+#'   field_types = field_types(
+#'     PrescriptionID = ft_uniqueidentifier(),
+#'     PrescriptionDate = ft_timepoint(),
+#'     AdmissionDate = ft_datetime(includes_time = FALSE),
+#'     Drug = ft_freetext(),
+#'     Dose = ft_numeric(),
+#'     DoseUnit = ft_categorical(),
+#'     PatientID = ft_ignore(),
+#'     Location = ft_categorical(aggregate_by_each_category = TRUE)
+#'   ),
+#'   override_column_names = FALSE,
+#'   na = c("", "NULL")
+#' )
+#'
+#' # aggregate the data
+#' aggregated_data <- aggregate_data(
+#'   source_data,
+#'   aggregation_timeunit = "day"
+#' )
+#'
+#' aggregated_data
+#' @seealso [prepare_data()], [report_data()]
+#' @export
+aggregate_data <- function(source_data,
+                           aggregation_timeunit = "day",
+                           show_progress = TRUE) {
+  # TODO: allow user to override existing aggregation_functions?
+  # TODO: raise an error/warning if data is less granular than aggregation_timeunit
+
+  log_function_start(match.call()[[1]])
+
+  validate_params_required(match.call())
+  validate_params_type(match.call(),
+    source_data = source_data,
+    aggregation_timeunit = aggregation_timeunit,
+    show_progress = show_progress
+  )
+
+  log_message(
+    paste0("Aggregating [", source_data$sourcename, "] by [", aggregation_timeunit, "]..."),
+    show_progress
+  )
+
+  # create column to group by, which will become the x-axes
+  timepoint_group_sequence <- create_timepoint_groups(
+    timepoint_field = source_data$data_fields[[source_data$timepoint_field_name]],
+    aggregation_timeunit = aggregation_timeunit
+  )
+
+  # get timepoint field values as they will be used repeatedly later
+  timepoint_field_values <- data_field_vector(source_data$data_fields[[source_data$timepoint_field_name]])
+
+  ### AGGREGATE OVERALL DATASET
+  log_message(paste0("Aggregating overall dataset..."), show_progress)
+  # load aggregated data into new vector
+  log_message(paste0("Aggregating each data_field in turn..."), show_progress)
+  agg_fields <- vector("list", source_data$cols_imported_n + 2)
[…]
