### 2022-10-14 745aba6
Minor tidying
