### 2022-10-15 8c8b29d
More styling. rOpenSci item (r2-14)
