+++ R/source_data.R
@@ -4,23 +4,22 @@
 # -----------------------------------------------------------------------------
 #' Prepare source data
 #'
-#' Validate a data frame against a [field_types()] specification, and
-#' prepare for aggregation.
+#' Validate a data frame against a [field_types()] specification, and prepare
+#' for aggregation.
 #'
 #' @param df A data frame
-#' @param field_types [field_types()] object specifying names and types
-#'   of fields (columns) in the supplied `df`. See also
-#'   [field_types_available].
+#' @param field_types [field_types()] object specifying names and types of
+#'   fields (columns) in the supplied `df`. See also [field_types_available].
 #' @param override_column_names If `FALSE`, column names in the supplied `df`
 #'   must match the names specified in `field_types` exactly. If `TRUE`, column
-#'   names in the supplied `df` will be replaced with the names specified in `field_types`.
-#'   The specification must therefore contain the columns in the correct order.
-#'   Default = `FALSE`
+#'   names in the supplied `df` will be replaced with the names specified in
+#'   `field_types`. The specification must therefore contain the columns in the
+#'   correct order. Default = `FALSE`
 #' @param na vector containing strings that should be interpreted as missing
 #'   values, Default = `c("","NA","NULL")`.
-#' @param dataset_description Short description of the dataset being checked. This
-#'   will appear on the report. If blank, the name of the data frame object will
-#'   be used
+#' @param dataset_description Short description of the dataset being checked.
+#'   This will appear on the report. If blank, the name of the data frame object
+#'   will be used
 #' @param show_progress Print progress to console. Default = `TRUE`
 #' @return A `daiquiri_source_data` object
 #' @examples
@@ -132,7 +131,7 @@ prepare_data <- function(df,
   }
 
   # Validate data against specification, store warnings instead of printing them
-  # use readr::type_convert for now.  Ideally want to store original values and describe action taken too
+  # use readr::type_convert for now. Ideally want to store original values and describe action taken too
 
   # ensure all columns are character type because readr::type_convert won't skip numeric columns
   dt_datatypes <- vapply(dt, typeof, character(1))
@@ -143,13 +142,13 @@ prepare_data <- function(df,
       data.table::data.table(
         col_index = which(
           dt_datatypes != "character" &
-            !vapply(field_types, is_field_type_ignore, logical(1))
+            !vapply(field_types, is_ft_ignore, logical(1))
         ),
         row_index = NA,
         message = paste0(
           "Data supplied as ",
           dt_datatypes[which(dt_datatypes != "character" &
-            !vapply(field_types, is_field_type_ignore, logical(1)))],
+            !vapply(field_types, is_ft_ignore, logical(1)))],
           " instead of character, non-conformant values will not be identified"
         )
       )
@@ -210,7 +209,7 @@ prepare_data <- function(df,
   # TODO: should I remove them here or when aggregating?  Summary doesn't look
   # right if remove them here. Rownumbers in warnings no longer matches either
   # TODO: check don't duplicate any messages from above
-  timepoint_index <- which(vapply(field_types, is_field_type_timepoint, logical(1)))
+  timepoint_index <- which(vapply(field_types, is_ft_timepoint, logical(1)))
   timepoint_field_name <- names(timepoint_index)
   if (anyNA(dt[[(timepoint_field_name)]])) {
[…]
