### 2022-10-15 86536cc
Make it clearer why users should use read_data() instead of their usual file reader. rOpenSci item (r2-23)
