### 2022-10-16 #535 @phuongquan
Dear @mbcann01 and @maurolepore,

Many thanks for your time and effort and patience in reviewing this package, I really appreciate it. Please see my responses to your comments below, with links to the github commit where relevant. (And apologies for some loss of formatting)

### Response to @mbcann01

The example in the "why should I use it" section of README. The figures are nice, but I'm not 100% sure what the connections I'm supposed to be making. Perhaps showing the Y-axis would help me to interpret the figures. Some users may even appreciate you spelling it out for them. Something like, "in the first figure... As you can see, that is reflected in the second figure as...".
**Response:** I have updated the text describing the figures and increased the size of the x-axis labels to try to make it clearer what is going on. (phuongquan/daiquiri@8b2c981ce39b092da9ae7f22a2301f6cd355eb66)

Example data. This is certainly optional, but I've started including example data in my packages. In the README, it looks as though you are including an example csv file in inst/extdata. To make testing/following along even easier for your users, you may want to consider converting this to exported data that users can load with data(daiquiri_data) or something like that. Here are some instructions if you are interested: https://r-pkgs.org/data.html#sec-data-data
**Response:** I have deliberately included a csv file rather than an R-specific format in order to be able to demonstrate all the functionality of the package, which includes reading in data from a csv.

I didn't see instructions for contributing anywhere on the README. I apologize if I just overlooked them.
**Response:** The contributing instructions are in the CONTRIBUTING file in the .github folder.

### Response to @maurolepore

o	[ ] (r02-01): In the vignette consider removing the literal reference to
"NOTE". The text itself is noteworthy and the announcement "NOTE" can be
dropped and yet be clear. Similarly, the headings "Example:" seems redundant
when it's followed by the text "For example, ...". You may drop the heading as
the sections are short and read well without subdivision.
**Response:** I have considered each appearance of "NOTE" in the vignette, and have removed/consolidated those instances where I feel they do not add anything, while I have left them in where I judge they serve to re-focus the attention of the reader (phuongquan/daiquiri@7fd31438b523efd206d4a08dd5b02497a7d8365a). I have left in the Example headings as I generally find it useful to highlight the locations of examples so that readers can easily find them when skimming the text.

o	[ ] (r2-02): Please ensure to document the return value of every function,
even it is NULL. CRAN enforces it. For example fix this in log_close().
**Response:** I have updated functions that previously returned NULL to return invisibly instead, and documented (phuongquan/daiquiri@b66ca9a6de8c7e7a6ab0238365825211e240b48a). I have updated `close_log()` to now return a path to the closed log file, and documented (phuongquan/daiquiri@f05f70baba4e75b458d9ea7abf5c8b43999cf7a3).

o	[ ] (r2-03): Ensure to format code and links as such.
With roxygen2 you do this with backticks and square brackets, respectively.
For example, the documentation of fieldtypes_template() mentions
"fieldtype()" -- which is not formatted as code nor linked to its helpfile.
**Response:** I have checked all documentation and updated code and link formats where found (phuongquan/daiquiri@ac191e13645a67d16a1afb4d5e7af196bd23286e)

o	[ ] (r2-04): The package documentation typically lives in R/packagename.R
not R/main.R. See ?usethis::use_package_doc() and
?usethis::use_import_from().
**Response:** I have moved the package documentation (phuongquan/daiquiri@7f52bca41df350afa4b98a79d7f499634029b7ff, phuongquan/daiquiri@0526c90db4dea40f372930fed63ee1945e880f79)

o	[ ] (r2-05): Consider using more of the convenient features of roxygen2,
e.g. use [some_link] instead of \link{some_link} and some_code instead of
\code{some_code}.
**Response:** I have enabled markdown and updated code and link formats where found (phuongquan/daiquiri@ac191e13645a67d16a1afb4d5e7af196bd23286e)

o	(r2-06): Ensure the use of \donttest{}, \dontrun{}, and \dontshow() as
per CRAN policies. They are common sources of rejection.
**Response:** I have spent quite some time reading about this and am still not 100% clear on what correct practice is. I have opted to use \donttest{} to wrap (unavoidably) long-running code as it has been suggested that \dontrun{} may cause problems with initial CRAN submission. (https://github.com/r-lib/devtools/issues/2216)

o	[ ] (r2-07): Whenever possible, print the result of the examples. For
example the result of fileldtypes() is assigned to fts but not printed.
Printed results help the reader see the output without the need of running the
code themselves. Also please show one example of a ft_*() function (e.g.
fs_simple()) so the reader can see its structure.
**Response:** I have done this (phuongquan/daiquiri@c3d5fb7a943ceb55aec0813fbeeae4ad07754f27)

o	[ ] (r2-08): Performance might be an issue. I see tests and examples that
take a long time to run. Consider using smaller data or removing performance
bottlenecks.
**Response:** It is the knitting of the rmarkdown document (which I have no control over) that takes up all the time for the main and reports tests so I do not see how I can make this any quicker unless I don't test the creation of the markdown document, which I cannot do as it is the main purpose of the package. User feedback I have received so far has indicated that performance is not a problem. Please also see my response to point (r2-09).

o	[ ] (r2-09): I fail to run test-sourcedata.R -- it takes too long or freezes.
After removing test-sourcedata.R I successfully run all tests but they still
take too long. The slower your tests, the less frequently you'll run them, and
the more your code becomes prone to errors. For companison run the tests
of other popular packages on CRAN.
  ℹ Testing daiquiri
  ✔ | F W S  OK | Context
  ✔ |       133 | aggregatedata [13.4s]
  ✔ |        15 | fieldtypes [1.0s]
  ✔ |        24 | main [240.8s]
  ✔ |         7 | reports [110.8s]
  ✔ |        45 | utilities [1.4s]
**Response:** I don't understand why the tests take so long when you run them, as they run in a fraction of the time for me (on a Windows laptop, Intel i7 processor, 16GB RAM), and sourcedata is not nearly the longest running one:
ℹ Testing daiquiri
✔ | F W S  OK | Context
✔ |       133 | aggregatedata [1.4s]
✔ |        15 | fieldtypes
✔ |        24 | main [39.2s]
✔ |         7 | reports [19.4s]
✔ |        38 | sourcedata [5.1s]
✔ |        45 | utilities [0.2s]
Even on an underpowered Windows laptop (Intel i5 processor, 8GB RAM) the tests only took the following amounts of time:
ℹ Testing daiquiri
✔ | F W S  OK | Context
✔ |       133 | aggregatedata [3.5s]
✔ |        15 | fieldtypes [0.2s]
✔ |        24 | main [64.6s]
✔ |         7 | reports [24.8s]
✔ |        38 | sourcedata [8.6s]
✔ |        45 | utilities [0.4s]
Unfortunately I don't have access to any other operating systems to be able to test them.
And as mentioned before, it is the knitting of the rmarkdown document that takes up all the time for the main and reports tests so I do not see how I can make this any quicker unless I don't test the creation of the markdown document, which I cannot do as it is the main purpose of the package. Please can you advise on anything further you expect me to do.

o	[ ] (r2-10) Adapt the tests or error-class so that each test is sensitive to
a specific error. I see a number of cases where different errors are tested
against the same class and no other property of the error.
**Response:** I have split tests that tested multiple behaviours into smaller units (phuongquan/daiquiri@e3b1f8254c4088ee3c268e031ca8ea100734bb84)

o	[ ] (r2-11): In general, try to use one test file for each function you test
(e.g. usually each exported function). This plays well with development
functions in the usethis package like use_test(), use_test(), and
rename_files(). Because testthat automatically prints the file name in the
test report, this approach allows you to not type the name of the function
inside the call to test_that() -- making tests easier to write and maintain.
```
 # good
 # tests/testthat/test-f.R
 test_that("works as expected", { code })

 # bad
 # tests/testthat/test-main.R

 test_that("f() works as expected", { code })
```
**Response:** I have followed the file structure recommended in https://style.tidyverse.org/tests.html, i.e.

> "The organisation of test files should match the organisation of R/ files: if a function lives in R/foofy.R, then its tests should live in tests/testthat/test-foofy.R."

 I have therefore chosen not to change this.

o	[ ] (r2-12): Make your tests granular enough so that a specific failure would
point to a specific expectation, and thus a specific piece of code. This will
make debugging easier. In validate_columnnames() I see a single call to
test_that() has multiple expectations, some expecting good behavior and
others expecting errors; they will be more informative if those different
expectations live in different tests.
**Response:** I have split tests that tested multiple behaviours into smaller units (phuongquan/daiquiri@e3b1f8254c4088ee3c268e031ca8ea100734bb84)

o	[ ] (r2-13): Please resolve the items marked with "X" by the command
#535 (comment)).
The section Editor-in-chief instructions state "Processing may not proceed
until the items marked with ✖️ have been resolved."
**Response:** I have re-run the bot and there are no more ✖️

o	[ ] (r2-14): Please note all results by the command @ropensci-review-bot check package and look for opportunities to improve the package. For example,
the goodpractice section reports that many lines are longer than the
recommended width of 80 characters. You may reflow lines relatively easy (and
semi-automatically) and improve readability by a lot.
**Response:** I really find the 80 character limit to be near impossible to achieve and applying styler did not reduce the number of these. I have manually re-flowed cases where I could see the benefit, but didn't do those which are only slightly over 80 characters. (phuongquan/daiquiri@8c8b29dcd67467e39c44808ab05fdedcb320e083)
Re the "Function names are duplicated in other packages" check, I am considering renaming `create_report()` to `daiquiri_report()`, which violates the verb_noun format but increases uniqueness. I also wonder whether `read_data()` would benefit from a suffix (e.g. `read_data_raw()`, `read_data_as_char()`, `read_data_unconverted()`, `read_data_plain()`) but cannot decide if any of these are actually any better. For the other functions, I cannot see how duplication can be avoided unless I prefix every exported function with the package name (which I find ugly and distracting). And even if it were possible to find globally unique yet concise and descriptive function names, there's nothing to stop someone else creating a function with that same name at a later date.

o	[ ] (r2-15): See rOpenSci's guidelines about code
style.
Using a style that is widely used in R and particularly in rOpenSci would be
more friendly for contributors.
For more information on how to style your code, name functions, and R scripts
inside the R/ folder, we recommend reading the code chapter in The R Packages
book. We recommend the styler package for automating part of the code styling.
We suggest reading the Tidyverse style guide.
For example, please use snake case consistently everywhere -- including the
names of functions, files, arguments, and variables:
 INCONSISTENT            CONSISTENT
 override_columnnames -> override_column_names
 showprogress         -> show_progress
 save_directory       -> save_directory
**Response:** I have run styler on the whole package (phuongquan/daiquiri@4aeb7a971b5c51119ff536cf20ce0666657274cb). I have also implemented snake case across the package (phuongquan/daiquiri@4c60f780ea414d751560ebc5ef5f7faedeab434f, phuongquan/daiquiri@628b19fa7762a06ded6c518f8ed12c491f18149e, phuongquan/daiquiri@f26164a79df3a8b77e519b9c184198f56d744562, phuongquan/daiquiri@1501580aa50787cc7c40f10db7f89335ab31f8b3, phuongquan/daiquiri@4d653ef25ec3eb2a70a53f8de91dd5d7f520f64f, phuongquan/daiquiri@244a156fb41fbdf516493a2e963f6bf0eb9c4926). I have also reorganised the functions within each file to put the exported ones first and added separators between functions (phuongquan/daiquiri@cb4c703dcdf95ff660f7c6d3071c5446abbb19fc).

o	(r2-17): See rOpenSci's guidelines about naming your
package. In a
Google seardch the name "daiquiri" seems be too common to help discover the
package.
**Response:** I have thought hard about the name and have selected it for phonetic reasons (**d**ai**q**ui**r**i as shorthand for **D**ata **Q**uality **R**eporting) and for ease of remembering it. When Googling "daiquiri package" it appears on the first page.

o	[ ] (r2-18): Ensure exported functions use the format verb_noun()
consistently. For example I see this format in create_report() but not in
log_close().
**Response:** I have renamed `log_close()` to `close_log()`, `log_initialise()` to `initialise_log()`, and `fieldtypes_template()` to `print_field_types_template()` (phuongquan/daiquiri@9ee037374ad4b537191580b4e7d702512d187c8e, phuongquan/daiquiri@52d302ba5786d85c81799956079262bdf0776345)

o	[ ] (r2-19): Use related prefixes consistently. For example fieldtype() and
fieldtype_template() are consistent, but ft_timepoint()is not. (But note thatfield_type*() is better because it's also consistent with other names
using snake case.)
**Response:** I have replaced `fieldtype()`/`fieldtypes()` with `field_type()`/`field_types()` (phuongquan/daiquiri@300ec048cb979aa99123342918733f622a290ae6). However, while I can see that `field_type_timepoint()` etc is more consistent than `ft_timepoint()` etc, I find the `ft_` prefix much more concise and easier to read. e.g.
```
fts <- fieldtypes(
	PatientID = ft_uniqueidentifier(),
	TestID = ft_ignore(),
	TestDate = ft_timepoint(),
	TestName = ft_categorical(aggregate_by_each_category = FALSE),
	TestResult = ft_numeric(),
	ResultDate = ft_datetime(),
	ResultComment = ft_freetext(),
	Location = ft_categorical())
```
versus:
```
fts <- fieldtypes(
	PatientID = field_type_uniqueidentifier(),
	TestID = field_type_ignore(),
	TestDate = field_type_timepoint(),
	TestName = field_type_categorical(aggregate_by_each_category = FALSE),
	TestResult = field_type_numeric(),
	ResultDate = field_type_datetime(),
	ResultComment = field_type_freetext(),
	Location = field_type_categorical())
```
I therefore have a preference to not change this unless it is considered absolutely necessary.

•	[ ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
**Response:** I have added both reviewers to the DESCRIPTION file. (phuongquan/daiquiri@d7511a1f6e02ef2510468bd1fef6c1d38666c7dc)

•	(r2-20) Recently I have seen a push towards quarto. Consider the pros and cons
of quarto versus rmarkdown.
**Response:** I have seen and read a little bit about quarto and my understanding is that currently it doesn't add very much to existing rmarkdown functionality but rather is more valuable for use with other languages such as python. The main drawback is that it requires a separate installation which can be problematic for users in tightly-controlled environments. I will bear it in mind for the future but at present I do not think that it is beneficial to change to using quarto.

•	[ ] (r2-21): Rename the example dataset more specifically to preserve the
option of adding more datasets in the future. For example, if later you have two
datasets, the names example_dataset.csv and example_dataset2.csv are less
descriptive than prescriptions.csv and surgeries.csv.
**Response:** I have renamed it to example_prescriptions.csv (phuongquan/daiquiri@08589156331b1aabb65630c580480f3bb13a69f8)

•	[ ] (r2-22): Remove unused code. For example, I see commented-out code in
tests/testthat/test-utilities.R and R/changepoints.R. Should you need that code
later, you can find it with git log -S "some code". And if you plan to work
with that code soon, it may be best to move it to a GitHub issue and follow up
in a GitHub pull requests when you are ready.
**Response:** Commented-out code has been removed (phuongquan/daiquiri@fc55cd9a72879f06fc9650c5ddd84e7fdb902176, phuongquan/daiquiri@db318706977e6e01ed40b97edba2a073f8e409e3, phuongquan/daiquiri@d986e84962d1dcf06cfe8266195dc904848a9cc4, phuongquan/daiquiri@e4f5b4e4310f5b18b3c55556d69093f9e787bee7, phuongquan/daiquiri@579fd90c1178775293c4326c93a5f56b96ff091d)

•	[ ] (r2-23): Consider removing the function read_data(). It seems to
be a thin wrapper around readr::read_delim(). If you keep it please do the
following:
1.	Explain why users should not read the data with readr::read_delim() and
friends directly.
2.	Consider preserving the name of the arguments to readr::read_delim()
and their original defaults to avoid confusing users of that popuolar
function.
3.	Allow users to pass arguments to readr::read_delim() via ....

**Response:** I created the `read_data()` function intentionally after user feedback showed that the default parameters used by `read_delim()` etc led to unintended behaviour (i.e. they interfere with daiquiri's ability to check for non-conformant values), so therefore I feel it is important to keep it.
1. I have reworded the documentation title and description to try to make the "why" clearer (phuongquan/daiquiri@86536cca198a100a48c869b7847bd91c04b73363)
2. I have prioritised naming consistency within the daiquiri package where there is a conflict, and otherwise have left the read_delim parameter names unchanged. The defaults have been deliberately changed to work well with daiquiri
3. While this initially looked like it would be a straightforward addition, it has turned out to be far more complicated than expected.
While I can see the benefit of allowing users to pass in additional parameters via `...`, I am unsure how to do this without opening the possibility of conflicts between the parameters I have specified deliberately and which should not be changed (otherwise it defeats the purpose of using the function), and the user potentially trying to (re-)set these parameters themselves. For example, I have preset `col_types` so that all columns are read in as character values, which is necessary for daiquiri to detect non-conformant values. If I leave this parameter out of the function definition (current behaviour), then the user might try to specify it via `...` . If I make the parameter available in the function definition, the user may try to update it. Either way, I would need to ignore what they pass in otherwise it defeats the purpose of using the function, but if I allow them to pass it in but then ignore it, that will be confusing to the user. I have come up with the following options:

Option 1: add custom parameter checking to ensure users don't pass in certain parameters (e.g. by raising an error if they try to pass in `col_types`, or by ignoring the parameter and returning a warning). If I return warnings I also have to remove the undesired params from `...` and then (presumably) use do.call to call read_delim() which adds complexity.

Option 2: don't include `...` and state in the documentation that the read_data() function is intended as a helper for beginner users, and if the user wants to use more advanced features of `read_delim()` then they should use `read_delim()` directly.

Option 3: include `...` but state in the documentation that any additional parameters are set at the user's own risk and may lead to unintended behaviour.

[…]
