### 2022-10-23 #535 @maurolepore
Thanks @phuongquan, your changes were made to my satisfaction and I recommend approving this package. Following are my responses to some of your comments. And after that is my approval via rOpenSci's template.

- r2-08: As long as you've considered this issue and understand, I'm happy.

You may place slow tests in tests/testthta/slow/ (or similar) so that your
frequent calls to `devtools::test()` ignore them. And you may run them
occasionally on demand with `testthat::test_dir(testthat::test_path("slow"))` (or a
`test_slow()` helper).

CRAN (or external collaborators) may be unwilling/unable to use as much
resources as you have. To test this package in a limited system you may
use a free project on https://rstudio.cloud/.

- r2-14: RE "Function names are duplicated in other packages", the bot
now says addressing this issue is optional.

- r2-23: As you considered this deeply, I'm confident you'll make a good
decision. Go ahead with what you think is more user friendly and easier
to maintain, or the best balance between the two.

Please remove me from DESCRIPTION. I added my agreement by accident. It's
out policy that editors shouldn't be acknowledged. Your contribution is
enough to make us all happy. Thank you!

## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 12
