### 2022-10-25 #535 @ropensci-review-bot
Approved! Thanks @phuongquan for submitting and @brad-cannell, @elinw for your reviews! :grin:
