### 2022-11-02 #535 @phuongquan
Hi @maurolepore, I have done everything except the central docs item, please see the comment below.

I found an issue with the rOpenSci central docs. I have a link to an example report in the menu bar which works in the github pages site https://ropensci.github.io/daiquiri/misc/example_data_report.html but returns a 404 error in the central docs https://docs.ropensci.org/daiquiri/misc/example_data_report.html. I have therefore kept the `pkgdown` site for the time being.
