### 2022-11-03 #535 @maurolepore
That's great progress, thanks!

A good place to host the "Example report" is under vignettes/ as you did with the "[Walkthrough for the daiquiri package]((https://docs.ropensci.org/daiquiri/articles/daiquiri.html)". You may place it there with `usethis::use_vignette()` or `usethis::use_article()`.
