### 2022-11-25 #535 @phuongquan
@maurolepore @mbcann01,

I'm currently in the process of preparing a JOSS paper, are you happy for me to acknowledge you as reviewers in it?
