### 2022-12-06 #535 @phuongquan
Hi @maurolepore @mbcann01,

I just wanted to ask again if you are happy to be acknowledged as reviewers in a JOSS paper? I'd like to try to submit it by the end of the week so a quick response either way would be much appreciated. I won't include you if I don't hear from you.

Thanks
