### 2022-12-06 #535 @maurolepore
Oh, sorry for the delay. Your contribution is enough, thanks! rOpenSci's guidlines is to not acknowledge editors -- which I think it applies here because I was both reviewer and editor.
